[FIX] mrp_workorder: validate worksheet QC without wizard
Issue:

Worksheet quality checks can not be validated if they bypass the Quality
wizzard, e.g. when when there is no notes on the QC and it is not
related to worksheet data's.

Steps to reproduce:

- Install `quality_control_worksheet`
- Create a BOM for a product with an operation + add instr (list icon):
    - Control per operation, type: worksheet, template: Quality issues
- Create ans confirm an MO for 1 unit of your product
- Go to the shopfloor and fill the worksheet > Save
> You are never able to validate the instruction and hence the WO
> Can never be marked as done.

Cause of the issue:

Since ee1913d7080f6aee06dac004d56203f2c09d4bad (saas-17.2) e.g the
change on if condition:
https://github.com/odoo/enterprise/blob/4238d6b071a57b3378968064ad322ea39a8b1294/quality_control_worksheet/static/src/views/quality_worksheet_fromview.js#L23-L28
the worksheet quality check is not marked as passed by the
`action_worksheet_check` when the record is saved if no
`quality_wizard_id` is set in the context. It is therefore expected to
mark the quality check as passed only if it set through the validate
button of the worksheet dialog. However, if the quality check does not
have specific data's such as notes or a worksheet url, the worksheet
dialog will be bypassed and there will not be a `quality_wizard_id` in
the context of the action returned by `this.openWorksheet`:
https://github.com/odoo/enterprise/blob/4238d6b071a57b3378968064ad322ea39a8b1294/mrp_workorder/static/src/mrp_display/mrp_display_record.js#L357-L362
https://github.com/odoo/enterprise/blob/0353a4c4307c2ac8d5a109a0e49be34a510e99ef/mrp_workorder/static/src/mrp_display/dialog/mrp_quality_check_confirmation_dialog.js#L82-L85
https://github.com/odoo/enterprise/blob/0353a4c4307c2ac8d5a109a0e49be34a510e99ef/quality_mrp_workorder_worksheet/models/quality.py#L25-L27
It is therefore impossible to validate the quality check by any mean.
Furthermore, the mark as done button of the WO is only visible once all
the QC are either "pass" or "fail".

Fix:

If we were to call the `action_open_quality_check_wizard` rather than
the `action_quality_worksheet` the same action would be used but with
an additional wizard would be in the context:
https://github.com/odoo/enterprise/blob/0353a4c4307c2ac8d5a109a0e49be34a510e99ef/quality_control_worksheet/models/quality.py#L53-L67
That way, if the `this.openWorksheet` is called without a worksheet
dialog we will always have a wizard to rely on for the
`action_worksheet_check` to be performed `onRecordSaved`.

Other issue:

In the bypass condition is wrongly set since the `operation_note` field
is used but is not defined on the `quality.check` model:
https://github.com/odoo/enterprise/blob/4238d6b071a57b3378968064ad322ea39a8b1294/mrp_workorder/static/src/mrp_display/mrp_display_record.js#L358
This condition should rather refer to the `note` field and the length
should be used to determine if the string is falsy...

opw-4347352 and opw-4354876

closes odoo/enterprise#75624

X-original-commit: e08e6f07768134ed181301f72f8a45b867897b77
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>