[FIX] documents: fix traceback after some action
How to reproduce:
- Install documents_account
- In Document App, folder “Finance”, select 2 pdf
- Click on Action -> Merge PDFs
- Remove the scissors to merge the 2 documents
- Click on “Create Vendor Bill”

You get the error: Cannot read properties of null (reading "querySelectorAll").

This is the most reliable way to reproduce the problem but that problem occurs
also in other circumstances, for example sometimes when moving to trash
documents after some other actions.

We solve the problem by returning an empty selection when there is no element.

Task-4381458

closes odoo/enterprise#75142

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>