[IMP] web: upload spreadsheet/cohort data as file instead of form field
This is related to the introduction of the limit
`max_form_memory_size` in odoo/odoo

The download button of the spreadsheet view takes the data currently
displayed and sends it with a POST to `/documents/xlsx`.

With the introduction of the limit in the above revision,
the data is now limited to 10 MB. Sending anything larger would
return an HTTP 413 Request Entity Too Large

Uploading the data as a file instead of a simple form field
allows to overcome that `max_form_memory_size` upload limit.

opw-4523014