[FIX] helpdesk: fix issue with onboarding tour
Previously, when a customer was selected or input was entered,
the onboarding tour failed to proceed to the next step.

With this fix, the tour will correctly advance when the user provides input.

Without demo data, the database does not have Deco.so i	so I select
the partner.

When the user does not provide the ticket name 'SAP is bad, paid by Odoo,'
the tour will not move to the next step because we had checked a fixed string.
Therefore, I have removed it and checked the kanban class instead.

task-4372470

closes odoo/enterprise#76147

X-original-commit: c1ef72b3f2a246147e6a05d9f5b6c99afe996252
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>
Signed-off-by: Prakash Prajapati (ppr) <ppr@odoo.com>