[IMP] l10n_be_coda: allow zero date in CODA file lines
Add check if date string is not equal to "000000" before parsing it, in record types that were missing this check.

Some banks provide CODA files with the date set to "000000" which caused the `_parse_bank_statement_file` method to raise an exception and fail to parse such files.

Using the `statement['date']` as a fallback in such case, as it is already used for `balance_start_date` and `transactionDate` fields.

task-4348031

closes odoo/enterprise#75233

X-original-commit: f92f42fe334bdfa8c211a0c5cb2b208ef20193ed
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>