[FIX] knowledge: fix dark mode and mobile UX
This commit fixes an issue with an icon in dark mode and some UX on mobile.
The `All changes saved` icon inside the topbar wasn't really visible in dark
mode. To fix this, an invert filter is applied to the img when Odoo is in dark
mode. This way instead of black the icon is white.

In mobile, there is no possibility to create articles below a specific route,
because the buttons are invisible. To fix, this we rendered the buttons always
visible when the user is on mobile.

task-4377610

closes odoo/enterprise#77028

X-original-commit: f2183da4c8c8b359576f767b9098c2d36c5de02e
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>