[FIX] data_merge: company field from _inherits model
In commit fb14e5793f924de21755a2b8e72ef9c651c9fb43, a way
to handle related company fields was added, however this
doesn't take into account the case where a model _inherits
from another model, in that case the field is `inherited`
and should be treated the same way as a stored field.

This commit makes it possible to find duplicates for models
with an inherited company field.

opw-4368443

closes odoo/enterprise#79239

X-original-commit: 70a4b8c449f43371c69ae7f2f3257f0c6ef31dc9
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>