[FIX] hr_expense_extract: prevent misdisplay of approved expense
Steps to reproduce:
- Open Expenses.
- Upload a sample expense pdf. The blue ribbon shows up saying “”
- Do not click on the refresh button. Instead, put in some value for the amount and submit and approve the expense.
- Go back to the expenses list view

Issue:
Notice this expense is at the top of the list and grayed out

Cause:
The order is determined by the extract_state_processed field:
https://github.com/odoo/enterprise/blob/ba06e216ab73c73e3a67fd917623c6ad357b869f/hr_expense_extract/models/hr_expense.py#L21

An expense is muted when this field is set to `True`:
https://github.com/odoo/enterprise/blob/1df090289f3c45c200d133734989a6d9a8073145/hr_recruitment_extract/views/hr_applicant_views.xml#L67

Solution:
We override `iap_extract` method
https://github.com/odoo/enterprise/blob/2e3113562309bf4d440e12ab09eaf94d0b302f1f/iap_extract/models/extract_mixin.py#L69
If the expense is not in draft, it indicates that the process continued without waiting for the extraction process to complete (values were manually entered).
To ensure a consistent extraction state, we extend the "Check OCR Status" logic to prevent an infinite waiting_extraction status.

opw-4239404

closes odoo/enterprise#75202

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>