[FIX] l10n_mx_edi: fix warning lxml warning in payment receipt
Steps to reproduce:
- Make sure that PyPDF2 is in version 1.26.0
- Install l10n_mx_edi
- Switch to a Mexican company
- Create an invoice:
  * Customer: [a Mexican customer]
  * Payment Way: [Effectivo]
  * An invoice line with a product having a UNSPSC Category
- Confirm the invoice
- Generate CFDI via "Send & Print" button
- Register a payment with "Efectivo" Payment Way
- Update Payments on the invoice
- Go to the payment
- Force CFDI
- Make sure the CFDI status of the payment is "Signed"
- Print "Payment Receipt"

Issue:
PyPDF2 raises a traceback:
"IndexError: list index out of range".

Cause:
PyPDF2 is executing this code to format a warning:
file = filename.replace("/", "\\").rsplit("\\", 1)[1]
where filename is something liked "<xxx>" (coming from the qweb compilation).
Therefore, there is no index 1.
The warning is coming from lxml and is due to the fact that we are evaluating
a xml value as it is in a qweb template:
"FutureWarning: Use specific 'len(elem)' or 'elem is not None' test instead."
And since this commit:
https://github.com/odoo/odoo/commit/fd67b56fc695526d09706d2b6218519ea255fc30
Python's "warning.py" is overridden in PyPDF2.

Solution:
Use "len(cfdi['cfdi_node'])" as advised to check if the xml node has children
in the payment receipt template.

opw-4359466
opw-4366016

closes odoo/enterprise#75467

X-original-commit: 07fae63e381dfa7a8035f67837f19d619600d434
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>