[FIX] spreadsheet_dashboard_account_accountant: use reference
The formula in B42 doesn't follow the pattern of other formulas in this
table.
This change isn't strictly necessary because C42 contains the exact
same string that was hardcoded ("income"), but at least now all
formulas are consistent, reducing the risk of future errors and
misunderstanding.

This issue was spotted using the new irregularity map feature (task 4357729)

Task: 4387232
Part-of: odoo/enterprise#75262
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>