[FIX] documents(,_spreadsheet): ensure shortcuts handler integrity
Shortcuts held incomplete values: file_size was missing (very
visible in list view), handler was missing for spreadsheets.

Note that a constraint for shortcut `handler` can only be
added in master, but it's not too bad as we can read
`shortcutTarget` to get the correct handler on opening the
document even if the preview is broken.

It also solves the issue where the list renderer is embedded
as `shortcutTarget` is not known when opening a dialog to
search for a document (reproduce: open `Parent folder`
"Search More" in "Configure"/"Details").

Task-4387563

Part-of: odoo/enterprise#75696
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>