[FIX] stock_barcode: don't fetch all product
In the `get_specific_barcode_data_batch` method, we convert the search
domain if we use GS1 nomenclature.
In some point, we have this code:
```python
converted_barcodes_domain = expression.OR([
    converted_barcodes_domain,
    [(barcode_field, 'ilike', barcode)]
])
```
The issue with that is `converted_barcodes_domain` is first an empty
list, so the call to `expression.OR` results into `[(1, '=', 1)]]`
domain. That means when we will search products with this domain, we
will get every products.

This fix ensures `converted_barcodes_domain` is not empty before to call
`OR`, otherwise we simply assign the converted domain to the variable
without the use of `OR`.

closes odoo/enterprise#75348

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>