[FIX] documents: prevent drag and drop on documents in the thrash folder
Steps to reproduce:

1. Create a new Workspace
2. Upload any amount of documents
3. Delete the workspace
4. Go to the Thrash workspace and try to move the
documents that used to be in the deleted workspace

Issue: When deleting a workspace, this one will
get archived, however the documents inside of it
will be moved to the Thrash and still be linked to
the archived workspace, so we will no longer have
access to that folder unless we restore it.

Then the problem arises when we try to do this
`foldersById[record.data.folder_id[0]]` since
foldersById doesn't seem to include archived
folders, it will be undefined so we won't be able
to read the `.has_write_access` raising a traceback

Solution: In this case the proposed solution is to
follow the behavior in 18.0 where we can not move
the documents that are attached to an archived
folder unless this one is restored. We get rid of
the traceback by checking if we have a folder
before directly tryng to read the `has_write_access`

Additionally we added an override for `_createCategoryTree`
for getting rid of the folders archived, this affects
the case where we directly archived a folder and try to move
the documents that were inside of it from the thrash.
(folder will sill be present in the model when retrieving it
with foldersById)

Backport of fac02ad43d523e55546ab21f3cc9a44db18d137b

opw-4357753

closes odoo/enterprise#75322

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>