[FIX] documents: avoid error when access_token no longer exists
The goal of this commit is to avoid getting an error message when using
the /documents/touch/ route for an access_token that no longer exists.
This bug was detected in the document_delete_tour where this route is
used when the document has already been deleted. This caused a
"missing record" message to appear. This commit fixes this behavior.

closes odoo/enterprise#75295

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>