[FIX] pos_order_tracking_display: differentiate notifiations topics
Reverting commits:
* https://github.com/odoo/enterprise/commit/4c6c902651ad48b1064f1eb4eb326fa773b6b098
* https://github.com/odoo/enterprise/commit/aa264a24091be85bb9983ef2c2aded1e505828b6

We found a way to be coherent with the behavior of the next versions. In the
next version the status screen can be used without being synchronized on orders.

The problem before the first fix was that the notifications were send to
all traking displays.

`bus_service.addChannel(pos_tracking_display-${session.preparation_display.access_token});`
The bus will listen to all topics broadcasted on the given channel.

`subscribe("NEW_ORDERS")` listens for any "NEW_ORDER" topi across ALL the
channels the bus is subscribed to. So if another channel also broadcast
"NEW_ORDERS", it will also execute.

We now use the `access_token` to create unique topics. Only notifications
explicitly addressed to this unique topics are now handled.

opw-4265987

closes odoo/enterprise#75399

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>