[FIX] account_avatax_stock: use the sales order warehouse when computing taxes
Steps to reproduce
==================

- Create a company in Canada with the canadian fiscal localization
- Configure the avatax sandbox credentials
- Create a new fiscal position in the USA that uses the avatax API
- Create a USD pricelist
- Create a warehouse in Cleveland Ohio
- Create a salesorder with an Ohio based customer and the USD pricelist
- Select a storable product with an invoicing policy on delivered
  quantity
- Set the avatax category of the product to "Hand tools"
- Select the warehouse just created on the sales order
- Click on the "Compute taxes" button
=> The tax is set to 1%
- Confirm the sales order
=> The tax is still set to 1%
- Validate the delivery and create + confirm a new invoice
=> The tax is now 2.25%

Cause of the issue
==================

When computing the addresses to send to Avatax, the stock move addresse
was used, but at this point, they have not yet been created

Solution
========

We now use the warehouse from the sales order. This is only a
best-effort estimation of where the products will ship from. It's
possible products ship from a different warehouse than warehouse_id. For
example, you can change the default delivery route to pull from a source
location in a different warehouse. However, this is probably not a
common configuration. In the end, the only way to know for sure is to
run the inventory-related code, which we want to avoid just for tax
computation. Finally, it's also possible for users to change the
warehouse manually after the order is confirmed. This can also
potentially change taxes, and there's no way to anticipate this.

The account.move code will continue to pull the warehouse from the
stock.move, because we know it will be created before the invoice. Note
that even on account.move the shipping address remains an estimation for
products with an “invoice on ordered” invoicing policy.

opw-4073701

closes odoo/enterprise#75655

Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>