[FIX] documents_project: archive folder on project unlink
Purpose
=======
When deleting a project, its documents are archived but not the documents folder.
Fixing that by archiving the folder if every projects it is related to has been deleted.

Specification
=============
As a folder is created for every new project, it seems logical to archive
the folder if the project is deleted. This prevents having "flying" folders,
i.e. empty project folders not being linked to any project.

Technical:
- As the same folder can be linked to multiple projects, only archiving the folder
if every projects it is related to has been deleted.
- Using sudo(False) because documents access rules can be slow.

Task-4262380

closes odoo/enterprise#72082

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>