[IMP] account_accountant: intensify hard lock date UX warning
This commit improves the UX design of the lock date wizard, by
intensifying the warning on the hard lock date field, showing them in an
`alert-danger`.

We want to make it even more explicit that the hard lock is
irreversible, concerns all users, given we already have feedbacks from
users who messed with it by setting it too early.

closes odoo/enterprise#75433

Task-id: 4391192
Signed-off-by: William André (wan) <wan@odoo.com>