[FIX] account_reports: use current rate to calculate initial balance
Steps to reproduce the issue:

1. Create two Companies, one in EUR, the other in USD
2. In EUR Company, create two Invoices with the following settings:
    - Total price of 1000
    - Invoice date: today and one month ago
    - You must be able to isolate them using Partner Ledger filters (not date)
3. In USD Company, add EUR Rates as follows:
    - Today: 1USD = 1EUR
    - One month ago: 1USD = 0.9EUR
4. In USD Company, open General Ledger and isolate the Invoices
5. First, set the timeframe from beginning last month to end of this month
6. The Invoice of last month is using the current rate, with an amount of 1000$
7. Second, set the timeframe to the current month only
8. The Initial Balance is using the previous rate, with an amount of 1111.11$

Explanation:

The initial balance is the balance of the company before the start of the currently reviewed timeframe. To calculate it, we must update the options to target the time before the current timeframe. Since we retrieve the rate used for this calculation using the updated options, we use the last rate before the current timeframe instead of the rate used in the current timeframe.

Fix reasoning:

Since every amount must be calculated using the current rate, we will retrieve it using the main options and not the updated ones.

opw-4299153

closes odoo/enterprise#75441

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>