[FIX] account_avatax: handle different taxes that map to the same name
In rare cases there can be more than one tax on the same line with the
same taxName and rate. These taxes are different e.g.:

```
Charleston Co Educational Capital Improvements Tax (Special) at 1%
Charleston Co Tt (Special) at 1%
```

But `find_or_create_tax()` maps both of them to SC SPECIAL 1%.

This doesn't cause issues for invoice_line_ids, those amounts are
correct. But the journal items are not because there's only one tax
journal item for this tax and its summary total gets overwritten
instead of added to. This leads to the total on the invoice not
matching the sum of debits/credits in the journal items tab.

Fix it by adding to the summary total instead of overwriting it.

PS. An alternative solution is to create unique tax names by
incorporating the jurisCode and possibly other relevant fields. This
approach would enhance the accuracy of tax representation on
invoices. However, historically, we have prioritized brevity over
accuracy [1].

[1] https://github.com/odoo/enterprise/pull/68549

closes odoo/enterprise#75464

Signed-off-by: William André (wan) <wan@odoo.com>