[FIX] spreadsheet_edition: allow base user to restore a spreadsheet
Base users do not have access to the model `spreadsheet.revision` which
is directly affected in a collaborative context. The usual flow consists
to give access to the revisions once we check the access rights on the
related spreadsheet. The code that handled the restoration of previous
versions did not follow this flow and base users would face access
errors.

Task-4397512

closes odoo/enterprise#76463

X-original-commit: db3f2abf68b2a652b996fa98410d7709ca63a84a
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>