[FIX] knowledge: save article name on record save
This commit fixes an issue with the knowledge controller when a record is saved
without a name. Usually when an article has no name we ensure that it has the
name when saving.
One case where this issue wasn't covered is when you save the changes done via
the save button.

Before, the name of the article wasn't updated if it was not yet set and when
they clicked on save.
Now, the title is updated according to the logic applied everywhere.

task-4398229

closes odoo/enterprise#75497

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>