[FIX] l10n_ke_hr_payroll: PAYE negative amount
Problem
----------
With low values of Gross, the PAYE can be negative.
The must be always positive or equals to 0

Objective
----------
Make the PAYE computation positive.
Check the PAYE computation.

Solution
----------
It will be now the maximum between the previous computation and 0. To be always >= 0.
Order the Gross Taxable with the sequence number.
PAYE computation is correct, check the task for more details.

task-4255918

closes odoo/enterprise#77902

X-original-commit: e2a80f12260b6c6e873c710e3c3618f8a43a94a7
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>