[FIX] delivery_sendcloud: avoid carrier UserError in batch picking validation
Issue:

Certain carrier's API processes validity checks and parcels generation
one picking at a time so that on our end we are forced to send requests
one at a time at the end of the the `_action_done`:
https://github.com/odoo/odoo/blob/e4f68fc6eb2a666a07b6627b9dc0ec575b79d1e3/addons/stock/models/stock_picking.py#L918
https://github.com/odoo/odoo/blob/e4f68fc6eb2a666a07b6627b9dc0ec575b79d1e3/addons/delivery/models/stock_picking.py#L192-L196
However, since a UserError of any of the picking will cause a rollback
of the entire batch on Odoo's side, pikcings might end up being
processed on the carrier side but non validated on Odoo's side during
batch validation.

Steps to reproduce:

- Configure sendcloud shipping method + website + payment providers
- In your sendcloud configuration select sendcloud shipping product to
  not be able to handle products with a weight exceeding 10 kg e.g;
  bpost @home (0-10kg).
- Create 2 storable products with respective weigth 1 and 100 kg.
- Publish them on the website for sale.
- Open a private window, make an order via the website for the 1kg
  product by filling your cart and use sendcloud as carrier.
- Repeat the operation with the other product to generate a seconde SO.
- Back to your main window > Inventory Overview > Delivery Orders
- Select both deliveries > Action > Validate (in batch)

> While the first delivery order was processed on carrier side
> (can be checked on sendcloud's website)
> the UserError of the second DO rollbacked its validation in Odoo.

Fix:

As such, and since the carrier requests needs to happen after we have
validated the pickings in Odoo, we should avoid any rollback of the
transaction once any requests has already been successfully treated on
the carrier side.

Note:

It is not possible restructure the code to use `senf.env.cr.savepoint`
in a loop in order to reset the validation of pickings that are not
valid on the carrier side because savepoints can't be used more than 64
times before the server crashes and hence can't be used in record loops.

opw-4357325

closes odoo/enterprise#76300

X-original-commit: 4e4c967b4cea59873ed4bdf8ab74c406638e9f46
Related: odoo/odoo#192033
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>