[FIX] pos_preparation_display: fix kitchen display sorting
In a POS restaurant order, when reducing the product quantity of an order line, the reduced display line in the preparation display was always the first one created, regardless of which order line was changed.

Steps to reproduce:

1.  In the POS restaurant, add a product with a quantity of 2 and an internal note. Confirm the order.
2.  Add another order with the same product and internal note but with a quantity of 5.
3.  In the kitchen display, change the stage of the first display line (quantity 2) from "To Cook" to "Ready."
4.  In the POS, reduce the quantity of the second order line (quantity 5) to 4.

Issue: The system reduces the quantity of the first display line (the one in the "Ready" stage).

Cause: In `_process_preparation_changes` of `pos_order.py` in the `pos_preparation_display` module, display lines are filtered and sorted based on their creation order. However, since there is no 1-to-1 relation between display lines and order lines, it is impossible to directly target specific display lines.

Fix: After this commit, display lines are now sorted based on their stage. Priority for quantity reduction is given to the line in the earliest stage. 

opw-4354093

closes odoo/enterprise#75565

Signed-off-by: David Monnom (moda) <moda@odoo.com>