[IMP] *: adapt to new name placeholder on moves
*: account_reports, l10n_ec_reports_ats, l10n_es_reports, l10n_uy_edi

Description of the issue this commit addresses:

The community PR linked to this one modified the naming behavior of moves and
therefore tests have broken. This PR adapts thoses tests to the new behavior.

Old vs. New naming behavior:

Let's consider an empty journal. In this journal three moves are added with
decreasing date (move1 is the most recent, move2 is in the middle and move3 is
the oldest). now we batch the moves and post them together.

Before, the first move of a journal would consume the first sequence number
upon its creation even in draft and subsequent moves in would be named "/".
When batched and posted afterwards, in the `_compute_name()` method, a
`.sorted()` would order the moves by date (account.move(3, 2, 1)) and would use
that order to name them. As the first sequence number is burned, move3 will use
the second number and move2 will use the third number in the sequence.
Resulting names: move1: 1, move2: 3 and move3: 2.
With x moves: move1: 1, move2: x, move3: x-1, move4: x-2, ...

Now, since no sequence number name is ever assigned to a draft entry, upon
their creation, all three moves are named "/". And then, when they are batched
and posted. they will be ordered 3,2,1 just like before but when naming move3,
the first sequence number isn't consumed anymore so it will be named 1, move2
will be name 2 and move1 will be name 3.
Resulting names: move1: 3, move2: 2 and move3: 1.
With x moves: move1: x, move2: x-1, move3: x-2, move4: x-3, ...

Desired behavior after this commit is merged:

Tests that were relying on the fact that the first move of a sequence uses the
first sequence number upon creation and not when posted are now adapted to take
into account that sequence number are consumed when posted.

In l10n_ec_reports_ats, moves without a latam document number are ignored for
the ATS export because missing that attribute means they don't have names which
makes them unfit for the export anyway.

task-4241510

closes odoo/enterprise#73837

Related: odoo/odoo#185326
Signed-off-by: William André (wan) <wan@odoo.com>