[FIX] documents: keep inspector tags on one line
When a tag has both a very long category and a very long tag name,
the delete icon becomes misaligned due to text overflow. This causes
the clickable area for the icon to shift away from where the icon is
visually displayed. As a result, users might think they cannot delete
the tag because clicking directly on the icon does not trigger the
delete action.

To fix this, we can apply `max-width: 2ch` to the delete icon. Since the
icon will always be one character's width, giving them 2ch width will be
enough to properly display it without taking more space than needed,
and ensures that the icon remains properly aligned with the tag,
regardless of text length.

This ch unit corresponds to be Equal to the “0” (ZERO, U+0030) glyph
found in the font used to render it.

opw-4403905

closes odoo/enterprise#75690

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>