[FIX] planning: fix traceback when there is no resource calendar
Currently a traceback is occurring when there is no
working time (calendar_id) in resource, while creating a planning slot.

To reproduce this issue:

1) Install planning
2) Open any planning slot and redirect to resource from slot
3) Remove the Working Time of that resource
4) Now try to create a new planning slot for that resource

Error:- "ValueError: Expected singleton: resource.calendar()"

Here `calendar_id` is not required in resource. So when the user removes the
calendar_id from the resource and tries to create a new planning slot,
it leads to the above traceback.

Because in the below line, we try to get calendar_d from the resource,
if the resource is available.
https://github.com/odoo/enterprise/blob/7847510dc6ad8dec73f01470714aacb0c3510769/planning/models/planning.py#L611-L612

sentry-6150920391

closes odoo/enterprise#75853

X-original-commit: 8cab165f6cdde2b790cac9ad19564cdeb4fbf9da
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>