[IMP] l10n_eu_oss_reports: adjust tests for new rates
Description of the issue this commit addresses:

The community commit linked to this one changed the rates of luxembourgian
taxes used in the oss fiscal positions. Because of that, some tests are failing.

Desired behavior after this commit is merged:

Said tests are not failing anymore.

task-4369658

closes odoo/enterprise#75716

X-original-commit: 3ff3d49662162bfd326b0b812cca7972f4ce51c8
Related: odoo/odoo#190816
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>