[FIX] account_reports: put account_code in first xlsx column
Partial revert of https://github.com/odoo/enterprise/pull/67111.
Some users rely on the structure of the export format (column order).
The order is most frequently code, then name. And code is a required
field. Those reasons made us want to re-adapt the behavior.

We also want the amounts to be formatted with thousands separator and
to always have 2 decimals (even if .00)

task-4399107

closes odoo/enterprise#75740

Signed-off-by: William André (wan) <wan@odoo.com>