[FIX] account_reports: put account_code in first xlsx column
Partial revert of https://github.com/odoo/enterprise/pull/67111.
Some users rely on the structure of the export format (column order).
The order is most frequently code, then name. And code is a required
field. Those reasons made us want to re-adapt the behavior.

We also want the amounts to be formatted with thousands separator and
to always have 2 decimals (even if .00)

task-4399107

closes odoo/enterprise#77808

X-original-commit: c4c1c40767b00e1a4e7fe71ec4cc899ebe688cd5
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>