[FIX] sign: set max canvas size for sign requests
Steps
-----
1. Sign > 'Upload a PDF & Sign'
2. Choose a PDF with a large amount of pages (typically 40 pages
should trigger the issue).
3. The PDF is not visible, the viewer is fully white.

Cause
-----
The `canvas` HTML element `#canvas_layer_0` used to display information
related to radio buttons crashes if its height is too important because
there are a lot of pages in the PDF.
There's no issue with the PDF itself, when dragging a sign item over the
document, the canvas is hidden and the PDF appears correctly.
The exact problematic value depends on browser and environment, see https://developer.mozilla.org/en-US/docs/Web/HTML/Element/canvas#maximum_canvas_size

Solution
-----
Use a maximum canvas size but keep the styling of the canvas element to
cover all the PDF viewer. This means a scaling factor needs to be applied
to the position of the sign items between which the lines are drawn.

opw-4407698

closes odoo/enterprise#75765

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>