[FIX] documents: fix attachment not pointing to the copied document
When copying a document, the copied attachment still points to the orginal
record:
- if it is a document with an attachment associated to an other record (ex.:
hr.expense). The attachment of the copied document is pointing to the same
record as the orginal document. Ex.: for an expense, the expense has now
2 attachments linked (one linked to the original document and one linked to the
copied one).
- if it is a pure document, the attachment points to the orginal document

We solve that problem here by always making the copied attachment to the copied
document.

How to reproduce case 1:
- Install documents_hr_expense
- Upload a document in the Document App
- Create an expense from that document
- Duplicate that document
- Open the expense App
- The expense related to the document has now 2 attachments instead of one

How to reproduce case 2:
- Install documents
- Log in as admin and open Documents App
- In Marketing -> Brand 1, duplicate "LA landscape.jpg"
- Move "LA landscape.jpg" to the Trash and delete it (from the Trash)
- Return to Marketing -> Brand 1

The copy has disappeared as well and shoudln't.

It means that deleting the original document deletes all copies. It works also
if we delete the copy (it deletes the original).

Task-4221258

Part-of: odoo/enterprise#77281
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>