[FIX] hr_con: fix activity when time off officer is not set
Steps:
- Install the hr_holiday module
- Configure Time Off Type with validation set to `Approved by Time Off Officer`
  or `Both`
- Create a Time Off request without setting a Time Off Officer.

Description of the issue/feature this PR addresses:
When the validation type is set to `hr` or `both` and no Time Off Officer is
configured, an activity is incorrectly generated upon creating a Time Off
request.

Fix:
This PR requires adding responsible_ids in the test cases. The changes made in
this task remove activities when responsible_ids is not set, causing the test
cases to fail. Adding responsible_ids will ensure the tests pass successfully.

task-4351688

closes odoo/enterprise#75828

Related: odoo/odoo#190193
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>