[FIX] timer,timesheet_grid: call get_server_time when needed
Before this commit, the `get_server_time` is called each time a timer
reactive is created in a component since the server offset is computed
thanks to that method is only stored in the timer reactive object.

This commit creates a timer service as before to be able to contain the
data fetched by the server to avoid having to fetch it again each time
the timer is running.

task-4361124

X-original-commit: 1157015d34daec8313512120aff54b73e2969fd7
Part-of: odoo/enterprise#75782
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>