[FIX] iot: calipers were missing read_once action
`quality_iot` listens continuously for measure updates. It also provides a
"take measure" button for devices that don't send automatically their measure.

On caliper devices, measures are sent continuously, so methods to read value once were
not implemented, resulting in a error if we tried to call them.

To avoid displaying an error popup, we implemented those actions on caliper drivers.

opw-4332476

closes odoo/enterprise#75847

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>