[IMP] knowledge: replace File by EmbeddedFile plugin
Due to changes in the /file command, in html field the result of such
command is a static blue box with a link to a file attachment.

In order to preserve the original behavior, the EmbeddedFile plugin is
available, which modifies the /file command so that it renders the
uploaded file as an embedded component.

This commit makes sure that in the Knowledge app the File plugin (which
produces the static file box) is replaced by the EmbeddedFile plugin
(which produces the embedded component).

task-3522395

Part-of: odoo/enterprise#75045
Related: odoo/odoo#189203
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>