[FIX] web_studio: reportEditor: do not create the studioview with translations
Before this commit the inheriting studio view (containing the user's customizations)
had its translations updated with its parent's translated terms.
This was because in the previous implementation we copied all the parent's views into the
inheriting, which not the case anymore since odoo/enterprise@acdca428d7543530177b7368ca9baddfb340e97c

This behavior has become problematic because of how translations work.
To illustrate, create a report, put in some text, save. Then, add some style to that text (making it bold for example)
The result is that in other language than in the base one, the report looks as if it weren't edited:
there is no style, no bold text.

This was due to multiple parts of the code:
1. translate.py doesn't take into account tags or style to produce terms, only the text
2. fields.py tries to match an old term with a new term at write time with some kind of similarity threshold
3. When getting a field's translated terms, we cannot know whether those terms come from the an actual translation in DB, or from
the base language's default

Now that customizations are stored in the form of an inheriting view, with xpaths, we can go back
to the standard odoo's behavior for translations: the inheriting view has its own terms, and must be translated
individually.
This commit is considered the lesser of two evils:
1. losing terms when editing
2. not modifying translations, even though the semantic structure has changed

opw-4389993

closes odoo/enterprise#75888

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>