[FIX] website_sale_subscription: fix product's variant price on configuration dialog
Steps to reproduce the issue:

1) Open the create new product page
2) From the "General Information" tab, set the "Product Type" to "Goods",
and set a "Sales Price"
3) Check the "Subscriptions"
4) From the "Recurring Prices" tab, create 2 different pricing rules, for
different variants (like one for size S, and one for size M)
5) From the "Sales" tab, set an "Optional Product"
6) On the website's product page, keep the default variant and click
"Add to cart", the modal "Configure your product" will open
7) Close the modal by clicking "X" or "Esc"
8) Choose another variant that results in a different price (from the prices
configured in step 4)
9) Click "Add to cart" again

-> Notice that the product's price is no longer the expected recurring price,
but it's the "Sales Price" configured in step 2.

Reason:
When opening the modal, we query the `data-subscription-plan-id`
attribute from the DOM using JQuery's `data` method. However, this
latter is not meant to querying data attributes from DOM. Its usage is
to actually store data associated with the JQuery objects it's associated
with. It worked the first time when we opened the modal, as the first call
for it will try to initialize the internal store, if empty, with the value of the
appropriate `data-*` attribute from DOM, but any subsequent updates
to the underlying `data-*` attribute won't be seen by a new call to this
function.

Fix:
Use `attr` instead of `data`, as the former reads directly from the DOM,
every time it's invoked.

Refs:
https://api.jquery.com/data/
https://api.jquery.com/data/#data-html5
https://api.jquery.com/attr/

opw-4369487

closes odoo/enterprise#75969

X-original-commit: 1c9fc3f322366c85140f22a9dc4157af7992830b
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Hadi El Yakhni (hael) <hael@odoo.com>