[FIX] web_studio: reportEditor recompose tcalls not tag t
have a t call node with another tag than "t". Edit the view

Before this commit, there was a crash when saving the changes. The helper
that recomposes the arch with the different parts inside the t-call did not
account for node not to be "t" tags.

After this commit, it works fine.

opw-4383844

closes odoo/enterprise#75905

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>