[FIX] l10n_mx_edi: link global invoice attachment and document
Steps to reproduce the issue:

1. Activate Mexican Localization and use Mexican Company
2. Create an Invoice and set CFDI to public to True
3. In list view, select Invoice
4. In Actions, Create Global Invoice and Confirm
5. Log as User without Admin Rights
6. Select Invoice
7. Receive access error:

>     Sorry, [user] doesn't have 'read' access to:
>     - Attachment (ir.attachment)

Explanation:

When creating a Global Invoice, `ir.attachment` is created as well and linked to the Global Invoice `l10n_mx_edi.document`, but if the link from document to attachment is complete, the opposite is not the case. `ir.attachment.res_id` has a value of 0, and, when checking access rights for `ir.attachment`, we will use the `_search` method in which one of three conditions, two of them having a `res_id` check, must be fulfilled for `ir.attachment` to be considered available to the user.
https://github.com/odoo/odoo/blob/afdfbc4041f167ba6d6ff6c17b8432eb531df6b1/odoo/addons/base/models/ir_attachment.py#L561-L570
A user that did not Create Global Invoice and without `base.group_system` in their `groups_id` fulfills none of those conditions.

Fix reasoning:

The issue lies more within `res_id` not being `l10n_mx_edi.document.id` rather than a flaw in `_search`. This change should only apply when creating a Global Invoice. When sending an Invoice to the CFDI through the regular `action_send_and_print` process, `ir.attachment` is linked to `account.move` instead.

opw-4365535

closes odoo/enterprise#75031

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>