[IMP] account_report: add currency code column in xlsx export
Issue:
When we export a report to xlsx containing some amounts in foreign
currency, those amounts are exported as values in the file, and loose
any kind of currency symbols. It then becomes impossible to know what
currency they were expressed in.

Solution:
When a line has an expression_label starting with '_currency_',
adds a new column right next to the associated column, with the currency
3 letters code as value. This solution can only be used on custom engine
as a line must be present in the report.

task-4011008

closes odoo/enterprise#76441

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>