[FIX] l10n_co_dian: prevent error when reload DIAN configuration
Currently, An error occurs when reloading the DIAN configuration in an account
journal to obtain the security token of a certificate, if no certificate is
available.

Step to produce:

- Install the 'l10n_co_dian' module.
- Navigate to Settings and click on Certificates under the Certificates and Keys
section.
- And delete a 'Demo DIAN Certificate'.
- Open any Journals and click on a 'Reload DIAN configuration'.

IndexError: tuple index out of range

The error occurs because the system attempts to access the certificate's token
from a company at [1], but it is not available.

Link [1]: https://github.com/odoo/enterprise/blob/a354a19060338084df35a14ed7f21defc3c05c11/l10n_co_dian/xml_utils.py#L123

To resolve this issue, Raise a validation error if a certificate is not found
when the system attempts to retrieve its security token.

Sentry-6165073693

closes odoo/enterprise#76052

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>