[FIX] sign: Fix canvas for irregular page sizes
Before this commit, we used the size of the first
page to determine the size of the canvas for
reportlab. This would cause issues if we uploaded
a PDF with any subsequent page larger than the
first as we would attempt to draw the signature
outside the canvas. Example PDF included in the
support ticket.
This seems to only be an issue with PyPDF2 2.12.1.

This commit fixes this issue by setting the
height and width of the canvas to the respective
maximum for all pages.

opw-4293390

closes odoo/enterprise#76529

X-original-commit: 002a8d094596084a9c2b81610564956d383586c2
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>