[FIX] account_followup: display correct status
Steps to reproduce:
- Create an invoice with a due date 20 days prior and an amount of $100
- Create a payment of 120$
- Create an invoice of 100$
- Reconcile the second invoice with the payment
- Go to the follow-up report

Issue:
The "In Need of Action" red banner does not appear, even though the overdue amount is positive

Cause:
The calculation does not account for the already allocated/reconciled payment amount.
In this case, the payment shows a balance of -$120, despite being partially reconciled.
-120 + 100 = -20 ≤ 0, which prevents the status from being set to "In Need of Action."

Solution:
Replace the balance-based calculation with one that uses the `amount_residual` field.

opw-4299170

closes odoo/enterprise#77679

X-original-commit: 1eaee394618d46cdb5bf116c9a2a1cb4cd79a68c
Signed-off-by: William André (wan) <wan@odoo.com>