[FIX] appointment: prevent access error in calendar sync
Before this commit, if a user was an attendee of an event linked to
an appointment type, but did not have access to the appointment (not
part of the staff_user_ids), an access error would occur when trying
to sync their calendar with Google or Microsoft. This error happened
inside the `_get_customer_description` function overridden in the
appointment module.

opw-4402859

closes odoo/enterprise#78110

X-original-commit: 6d8ab0e04b0be5feb931be7679664eeb3aeb9d21
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>