[FIX] account_budget: prevent zero division error on entering invalid end date
This error occurs when the user enters an 'End Date' that is one day earlier
then the 'Start Date'(e.g. Start Date: '25/12/2024' and End Date: '24/12/2024').

Steps to reproduce:
- Install the 'account_budget' module
- Create a new budget, fill in all required fields, and add a line
- Start Date: '25/12/2024' and End Date: '24/12/2024'

Traceback:
ZeroDivisionError: float division by zero

At [1], this error occurs because when the 'End Date' is set to one day before
the 'Start Date,' the 'line_timedelta' becomes zero. As a result, the
denominator evaluates to zero.

[1]- https://github.com/odoo/enterprise/blob/3cfb6517220ae252d633bb9debc5010a0f14efbe/account_budget/models/budget_line.py#L77-L79

This commit resolves the error by adding a constraint to check for
this specific scenario, ensuring that line_timedelta is not zero.

sentry-6174377667

closes odoo/enterprise#76617

X-original-commit: 293d664369056935ed8cabcdaf204460d03521fc
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>