[FIX] documents: fix creating and deleting shortcuts
Reproduce:
1. Create a shortcut for a folder.
2. Move the shortcut to the trash.
3. Go to the trash and try deleting the shortcut.
4. You get a 'Missing Record' message.

This occurred because e1714438 introduced a bug that
creates attachments for shortcuts.
(The mimetype is a related field so it cannot be "synced" with the target
as it is not worth creating attachments for this.)

Also, removing two useless tests setups (feature change) and
completing the document unlink test for res_model `False`.

Task-4433373

Part-of: odoo/enterprise#76220
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>