[FIX] documents_spreadsheet: Fix template creation with grouped dates
When creating a spreadsheet.template, we ensure the values of each
groupbys are correctly ordered. Tthe current implementation assumes
that the groupby will return values present in the tables and not
values altered by a groupby with specific granularity (namely date
fields) by adding the values to the search domain. Unfortunately,
the search does not support dates with a granularity different than
`day`.
This revision ensures that we do not add such grouped values to the
search domain.

Task-4438631

closes odoo/enterprise#76298

X-original-commit: 11dfe2455173a158b1c1421012867fe5a38967a7
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>