[FIX] sale_subscription_stock: test's start date issue
Before this commit, the test was relying in "today's" date for
creating the recurring invoices. Since the time period for
generating them was too short, the expected deliveries were
not being generated.

After this commit, we give a fixed date to the start date
of the subscriptions and then generate the recurring invoices
with a time distance of a month, solving the issue.

Issue-from: OPW-4166852
Related PR: odoo/enterprise#73397

closes odoo/enterprise#78294

X-original-commit: df9d574b639d012e31b4e8441318ea64988c83c8
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>