[FIX] sale_renting: remove undeterministic step
In this commit, we remove the step concerning the external button
(which is visible when editing the input or hovering over it).
Depending on the speed at which we check the trigger, the step passes
(or not). For it to pass every time, we would have to hover over the
input and check that the external button is present.
This step was added when we removed the extra_trigger from the tours API.
In any case, it seems doubtful that the tour tests that this trigger is
in the DOM. We can therefore remove this step.

closes odoo/enterprise#76385

Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>