[FIX] social_push_notifications: disable push notifications when dependencies fail to load
When enabling push notifications on their website, the system injects a
widget that is responsible for (1) obtaining the user's permission to
receive popup notifications, (2) loading the Firebase dependencies, and
(3) initializing the Firebase app to handle push notifications.

Currently, the Firebase dependencies are loaded from the Google CDN.
Occasionally, these dependencies may fail to load, which can occur if the
remote server is unreachable or if the user has installed a privacy-oriented
add-on that blocks requests to Google. When the dependencies fail to load,
the widget responsible for push notifications crashes, and an intrusive
error dialog appears.

To avoid showing an error dialog on page load, we will silence the error
and disable the push notification feature if the browser fails to load
the Firebase dependencies.

task-4247811

closes odoo/enterprise#76409

X-original-commit: 5c1d464871f0f67e0f84c1c80c32e69cb098effc
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>