[FIX] l10n_ro_saft: fix test asset sale date
the setup of test class TestRoSaftReportAssets is no longer passing since 2024-12-31
because of an old unnecessary freeze_time.
This commit also fixes a second issue related to the sale of the asset. The value was
wrongly computed, and should be a fixed amount in order to make the test robust.

runbot issue: 111360

closes odoo/enterprise#76414

X-original-commit: 92e202b83cd2883d93cadfe8ecc73a99d6d7a5c0
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>