[FIX] documents_spreadsheet: disable public download button
Steps to reproduce:

- install Documents
- create a new spreadsheet
- hit the Share button
- change the sharing settings to "Anyone with the link: Viewer / Musty have the link to access"
- open the sharing link in an incognito window
- hit the download button

=> 404 not found

A public user cannot convert a spreadsheet to an excel file and downlaod it because a route needs to be called to create the excel file. This route is currently not allowed for public users.

For frozen spreadsheet, we generate the excel file before-hand (when the internal user shares the spreadsheet). For a regular "live" share though, we cannot pre-generate the excel file since the spreadsheet continues to live and evolve.

The `spreadsheet` module needs to be updated for this bug fix to take effect. (otherwise the download button won't do anything, which is not worse than facing a 404 response)

See also community PR.

closes odoo/enterprise#76460

Task: 4440241
Opw: 4417519
Related: odoo/odoo#192349
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>