[FIX] sale_stock_renting: MTO products
Steps:
- enable rental transfers
- activate MTO route
- create a product with MTO route and 0 stock
- create a rental order for that product, validate

Issue:
The rental order will create 2 pickings that are waiting for each other.
This is because the MTO move is using the same rental route with the stock
destination, so it ends on the return rental picking.

Fix:
Force the procurement to use the `route_id` of the `rule_id` of the origin move.
This way it will correctly create a new picking from Vendor to Stock for the product.
If the product has a Buy or Manufacture route, it will use those instead.

Also found 2 other issues:
- when enabling, then disabling, then re-enabling rental pickings, the stock
rule stays archived. It is now fixed and a test has been added.
- if you install Rental first, then install purchase or mrp without enabling
the rental pickings, Odoo will still create the rental return rule but without
a source location `rental_loc_id`. This is never added back again later.
The test `test_lot_accuracy_in_schedule` was adapted to fit this change.

closes odoo/enterprise#78699

X-original-commit: 0ef39a8aa6d663b0c46b3d8af4e60fc25c53859a
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Stéphane Diez (snd) <snd@odoo.com>