[FIX] l10n_mx_edi: prevent error when creating global invoice
When a customer creates a global invoice, a traceback will appear.

Steps to reproduce the error:
- Install "l10n_mx_edi" module > Switch to Mexican company
- Create an invoice without customer and product > Save
- Go back to List view of invoices
- Select that invoice > Actions > Create Global Invoice

Error: A traceback appears:
"TypeError: sequence item 0: expected str instance, bool found"

https://github.com/odoo/enterprise/blob/01f749f40fc4b5f9708696dc42eab7125e790b28/l10n_mx_edi/models/account_move.py#L306
When a customer creates a new invoice, "name" will be False,
When the customer tries to create global invoice,
So it will lead to the above traceback.

sentry-6197401330

closes odoo/enterprise#76515

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>