[FIX] l10n_it_pos: force load the current company's partner
Before this commit, it could happen that the partner associated
to the current company was not automatically loaded at the start
of the POS session.
Since this parter is required for proper operation of l10n_it_pos,
it could lead to unexpected errors.

After this commit, the partner is explicitly force-loaded at the
start of the POS session to ensure proper operation.

opw-4394672

Part-of: odoo/enterprise#76517
Related: odoo/odoo#193123
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>