[FIX] account_reports: gl print with alpha characters in search bar result in empty pdf
Before this commit, when we did a search in the general ledger for an account code with alpha characters, (ex: `411EDP` in the search bar) the options resulted with `411edp`.
It was changed to lowercase. And due to a change in the _name_search of account_account, it now uses like and not anymore ilike.
Commit: https://github.com/odoo/odoo/commit/1f589682a9414312b8c15beec5a09e76b16c7c41

The fix is to provide to the options the original search query without lowercase.

opw-4388517

closes odoo/enterprise#76796

X-original-commit: 11253f9e5d03fed74cf4f112dc236a01c7440721
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>