[FIX] account_reports: Tax > Account and Account > Tax Tax Reports: properly group results from the tax details query
To reproduce the issue:

1) Create a vendor bill for 100€, with a tax of 15%
2) In the grouped generic tax reports, the base will be 100 and the tax 15
3) Create an expense of 200€, with the same tax, paid by the company. Fully validate it and post the entry.
====> The grouped generic tax reports now should display 300 of base, and 45 of tax. It does only show 200 of base, though.

This happens because when post-processing the results of the query ran by the report, we do this https://github.com/odoo/enterprise/blame/16.0/account_reports/models/account_generic_tax_report.py#L406 , so we essentially don't add the base amounts of the tax lines whose identification key has already been treated. This is important for taxes with multiple tax repartition lines.

In our case, a line is generated by the query for both the invoice and the expense entry. This is because the query groups on display_type: the expense line has display_type = 'product', while the invoice has 'tax'. This grouping was introduced here https://github.com/odoo/enterprise/commit/67f841ee36cb5a8ef95ae3dd3859239a4d10b99d to avoid counting the base amounts of the cash rounding lines as additional base amounts.

We fix this by slightly tweaking the query, in order for rounding lines to be considered as having a base amount of 0.

opw-3890736

closes odoo/enterprise#76801

X-original-commit: 0a18d0244f658fe1a6fd8e3d5cb703e728a426f5
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>