[FIX] account_reports: manage additional balance column
Since 18.0^[1], the ending balance is done by summing the values fetched
instead of querying the database again.
While this is a great performance gain, it also means that we can't get
other values like the balance.
Note that the Trial Balance has been refactored in 18.1^[2], so a simple
fix is enough.

opw-4435450

[1]: e598fcb48b5e4f0126406a4008f175a88528ba85
[2]: a7e1ec20e07efc39bca44d3ae613a54770175273

closes odoo/enterprise#76535

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>