[FIX] l10n_uk_reports_cis: access into empty partner_id
Steps to reproduce:

1. Install l10n_uk_reports_cis
2. Open a new bill
3. Set the 20% G tax on a new line.
4. Save without assigning a partner.
5. A traceback is raised.

Description of the issue this commit addresses:

In the l10n_uk_reports_cis module, a tax validation system is implemented that
reaches into move.partner_id to check the chosen taxes validity. This causes an
issue when the move doesn't have a partner yet, raising a traceback.

Desired behavior after this commit is merged:

No traceback is raised.

Note on the fix:

An additionnal check verifying the move has a partner is
done before reaching into its values.

opw-4365356

closes odoo/enterprise#76527

Signed-off-by: William André (wan) <wan@odoo.com>