[FIX] account_iso20022: provide the BIC when possible even in CH
The fix previously done[^1] was always removing the BIC number because it
is "optional". While it should be optional, some banks require it for
payments done in EUR.
The reason the previous fix was done was to avoid having the `Othr` tag,
but that tag won't be used outside of methods returning `False`
explicitly for `_skip_CdtrAgt`.

opw-4423834

[^1]: 1a221f2f3160a48dc7ba08ea59c6781071b1adf3

closes odoo/enterprise#76558

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>