[FIX] helpdesk: team creation uses ASCII in aliases fallback
As per (ref.1) and consequently (ref.2), the mail.alias creation is not handling
characters outside of the ASCII set. This collides with the helpdesk team creation,
which uses the company name (which allows non-ASCII characters) to create an alias.

Since SMTPUTF8 is slowly spreading its reach, UTF-8 will probably be adopted.
This commit introduces a fallback mechanism for that case.

Reproduce
---
1. -i helpdesk
2. Attempt to create two companies:
3. Create a company with a name composed of two non-ASCII symbols (木日).
4.Create a company with a name composed of two non-ASCII symbols (日木).
BUG: can't create alias, "??" alias already exists!

References
---
(ref.1)
https://github.com/odoo/odoo/issues/53403#issuecomment-654334095

(ref.2)
[FIX] mail: prevent international char in alias
https://github.com/odoo/odoo/commit/6c154ab5f5af48480da6b77299687d9a768cd14f

opw-4420662

closes odoo/enterprise#76564

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>