[FIX] quality_{control_worksheet,mrp_workorder_worksheet}: let worksheet check fail in backend
**Current behavior:**
Performing worksheet quality checks in the backend for MO
operations which have workorders will not work- that is the
worksheet can be filled in but will be automatically passed
without actually respecting the conditions set on the quality
point.

**Expected behavior:**
The check should pass/fail depending on the conditions on the
point and how the worksheet is filled in.

**Steps to reproduce:**
1. Create a product and give it a BoM with an operation

2. Create a quality point like:
* `product_ids=[product from step 1]`
* `operation_id=Manufacturing`
* `measure_on='operation'`
* `measure_frequency_type='all'`
* `test_type_id=worksheet`
* `worksheet_success_conditions=[('x_passed', '=', True)]`

3. Create an MO for the product, confirm

4. Go to the quality point and open the new check via smart
button

5. Open the worksheet, see `Passed=False`, save

6. Click the `Check Worksheet` button -> see that the check
passes despite the requisite field not being set

**Cause of the issue:**
When performing the check in this manner, because it's not
really the "expected" sequence of actions, the method will reach
https://github.com/odoo/enterprise/blob/8a120a207e283872a6296437236c409c9a4f3cef/mrp_workorder/models/quality.py#L550
and be passed, never checking the success condition domain
against its worksheet.

**Fix:**
Create a quality check wizard in the event that one doesn't
exist and we aren't in the act of filling out the worksheet
(`not self.env.context.get('from_worksheet')`).

Then instead of returning the window close action and
inadvertently passing, the check is actually performed.

opw-4321840

closes odoo/enterprise#75711

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>