[FIX] web_studio: viewEditor: invisible fields are draggable
Before this commit, fields that were always invisible (with an `invisible` attribute equals to True|1)
were not listed in the Sidebar's "Existing Fields", and were not available for drag/drop.

This was rather impractical as such field may be in the arch just to have their data loaded, not really displayed.
Moreover, since odoo/odoo@6f06420e4a9443c52dc0cb427f8f55eb4aecabce, fields may be added in the view automatically
based on whether their value will be used in some condition or context elsewhere in the view.

After this commit, always invisible fields are always proposed for addition in the view.

opw-4450351

closes odoo/enterprise#76608

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>