[FIX] l10n_co_edi_website_sale: display Identification Number in billing address form for Colombia
Following a discussion with the product owner, it was clarified that in Colombia,
customers must provide both an Identification Type and an Identification Number (VAT) for electronic invoicing,
as these details are mandatory for all sales.
Additionally, for extra billing addresses, the fields should also be visible and required.

To address this, all billing addresses in the Colombian localization must include these fields as required.

1. Install the module "l10n_co_edi_website_sale."
2. Change the default company of the current user to a Colombian company and switch to it.
3. Go to the partner record for the user and change the address to a Colombian address.
4. Add a contact as an invoice address with a Colombian address.
5. Go to the shop, add a product to the cart, and click "Go to checkout."
6. On the address page, click "Save address." The page redirects to itself repeatedly.

opw-4309448

closes odoo/enterprise#76609

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>