[IMP] website_appointment: support turnstile
The previous commit implements captcha for this form

Turnstile can be implemented on top of it by just invoking the turnstile script in js.

As it's implicitly checked when recaptcha is checked, this also avoids any issue with the form potentially not working if turnstile is enabled.

task-4335141

closes odoo/enterprise#76672

Related: odoo/odoo#192827
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>