[FIX] appointment: resource_id saved on create event from calendar
Steps to reproduce:

1. Go to appointments
2. Select any appointment with resources.
3. Change from the gantt view to the calendar view.
4. Create a new event from the calendar view with
resource_id set.

Issue: Since in this form we are using the `resource_id``
and not the `appointment_resource_id` the inverse computation from
`_inverse_appointment_resource_id_or_capacity` will think that
we have removed the linked resource_id from the appointment so we will
remove the booking line.

Solution: In the form, we have 'appointment_resource_id' as invisible,
however, unless we set it to readonly it's going to be incorrectly
calling the `_inverse_appointment_resource_id_or_capacity`. We don't
need to save 'appointment_resource_id' here.

opw-4440569

closes odoo/enterprise#76653

Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>