[FIX] sale_subscription: recurring_total rounding
Subscriptions with multiple lines and tax-included prices introduced a
rounding error in some cases.

This commit fixes the _compute_recurring_total method by using the same
rounding method as _compute_amounts in the sale/ module, based on
AccountTax helper methods.

A new test is introduced :TestSubscription.test_recurring_total. The
issue was reproduced on any company and CoA but only for some foreign
currencies such as Bulgarian. Before the fix, this test would fail with:
- `non_recurring_total` = 0.01 (correct value = 0.00)
- `recurring_total` = 195.79 (correct value = 195.78)

opw-4460812

closes odoo/enterprise#76686

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>