[FIX] account_reports: fix traceback when installing inventory
Currently, a traceback occurs if the user installs inventory
module, when the accounting module was installed and
having no state_id in the company.

To reproduce this issue:

1) Install `Accounting` and remove the state from the company
2) Update the Fiscal Position record with Foreign Tax and Federal State
3) Now install the inventory module

Error:-
"TypeError: sequence item 0: expected str instance, bool found"

State_id is not a required field in the company. So can remove it
and when there is no state_id and a fiscal position record having a
foreign tax we get a traceback from below lines.

https://github.com/odoo/enterprise/blob/1c6f0273e76b6dfb89ef65eb359e620b9c4c5759/account_reports/models/res_company.py#L216-L219

This is because we have a `vat_fpos_with_state_count` value because of
a foreign vat in fiscal position, but we don't have a state_id in company.

So the `state_codes` value will be a list of bool, which leads to a traceback.

We can resolve this issue by making the code more robust.

sentry-6203209197

closes odoo/enterprise#76784

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>