[FIX] documents: in the list view, the parent folder is "Unnamed" in the inspector
Bug
===
If we are in the list view, and we are inside a folder, when selecting a document, it
will show "Unnamed" for the folder. The reason is that with `column_invisible=1`, the
display name is not loaded. To solve that, we can put the field in `optional="hide"`.

Task-4461038

Part-of: odoo/enterprise#76699
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>