[FIX] Payment_provider: Fix Company Assignment for Mandates
Behavior Before the Commit:
When generating a new payment link, the system creates a new mandate without explicitly setting the `company_id`. As a result, the default company (`company_id` of the public user, `user_id = 4`) is assigned, which may not match the intended company if the payment is being processed for a company other than the default one. This mismatch between the journal's company and the mandate's `company_id` prevents the payment from being completed.

Fix:
The `company_id` is properly set during mandate creation, aligning it with the payment link's associated company and preventing unexpected behavior.

Steps:
1) Configure two companies with accounting setups.
2) Open the Payment Providers menu.
3) Select SEPA, activate test mode, and ensure it is published.
4) Duplicate the SEPA payment provider, set the company_id to the second company, activate test mode, and ensure it is published.
5) Go to any sales order in the first company, generate a payment link, open it, and select SEPA. Enter any fake IBAN — it will process successfully.
6) Switch to the second company and repeat step 5. You will encounter an error message because the mandate is incorrectly assigned to the first company, preventing payment completion.

closes odoo/enterprise#76827

X-original-commit: b0146e8933e647da3138e9fb90c2efb4b5af7efa
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>