[FIX] knowledge: replace composer content with clipboard text
Recent updates introduced automatic inclusion of the user's signature to
the mail composer (see: odoo/odoo#184695). When the user clicks on the
"Send as Message" button of a clipboard block in Knowledge, the macro
system will do a few steps to (1) restore the previous view, (2) open the
full mail composer and (3) insert the clipboard text at the end of it.

As the signature is added when opening the full mail composer, the
clipboard text is inserted after the user's signature which is
inconvenient as users must manually adjust the message to position the
signature at the end of the message.

To avoid this problem, we'll replace the full mail composer content with
the concatenation of the clipboard text and the user's signature. This
guarantee that the two elements will appear in the desired order.

Steps to reproduce:
1. Open the CRM lead view
2. Press `Ctrl` + `k` to open the command palette
3. In the command palette, select any article
4. In the article you chose, insert a clipboard block with the /clipboard command
5. Click on the "Send As Message" button of the clipboard
=> The macro restores the CRM lead view, opens the full mail composer and
   paste the clipboard content after the user's signature.

TO BE: The macro should replace the message of the full composer with
the clipboard content and the user's signature should be placed after
the clipboard content.

Task-4428445

closes odoo/enterprise#76725

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>