[FIX] stock_barcode: take selected line into account when scanning
Steps to reproduce:
- Install stock_barcode and Sales
- In Inventory settings, enable "Multi-Step Routes"
- Go to "Inventory / Configuration / Warehouse Management / Routes"
- Configure the rule of "Deliver in 1 step (ship)":
  * Propagation of Procurement Group: Fixed
  * Fixed Procurement Group: [create a new one]
- Create a product tracked by lot
- Update the quantity (e.g. 100) of the created product
and assign it to a lot
- Create a SO for a customer with the created product
- Confirm the SO
- Create another SO for the same customer with the same product
=> A delivery order should be created with 2 operations with the
same product
- Open the delivery order in Barcode
- Scan the source location
- Select the 2nd line
- Scan the lot

Issue:
The first line is incremented instead of the second one.
Even when all the products for the first line will be scanned, it will
not be possible to scan the lot to increment the second line.

Cause:
The selected line is not taken into account when searching the line
on which the scanned lot should be applied.

Solution:
Unshift the selected line to the first position of the array of lines
to prioritize it when searching for the adequate line for a scanned lot.

opw-4208612

closes odoo/enterprise#77700

X-original-commit: 3613bad4ff5aa711fd0cc24b36437d9952ff51ce
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>