[FIX] knowledge: fix comments popover display when top position is 0
Purpose:
--------
The comments popover is not displayed when the computed popover top value
is "0". This issue arises from incorrectly using this value as a boolean
condition, whereas "0" is an acceptable top value (= top of the body).

To fix this issue, we now check if the popover's top is undefined instead
of using it as a boolean.

Task-4461321

closes odoo/enterprise#76703

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>