[FIX] knowledge: fix disappearing comments overlay
This commit fixes an issue with the comments overlay when zooming and dezooming
on the article.
At certain sizes the overlay of the comment would disappear without reason.
This is caused by the rect computations done to identify the correct overlay to
draw.
The target selected by the coordinates given to the `elementFromPoint` method
would be incorrect.
This happens because the coordinates weren't the center of the rectangle but the
top. Which means that for certain resolutions the element selected is the whole
body instead of the correct paragraph/text node.

Now, the coordinates are computed so that we are indeed at the center of the
rect.

task-4463401

Part-of: odoo/enterprise#76709
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>