[FIX] social: text not truncating on social post links
Before this commit:
Post including long links were not truncated and will extend
beyond the kanban card.
This happen as the display CSS property is set to `inline` by
default due to the <field> tag being transformed into a <span> tag
in the HTML. <div> default display is `block` while <span> are
`inline`.

After this commit:
The link truncate as intended with "..."

opw-4446906

closes odoo/enterprise#76719

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>