[FIX] web_studio: edit default value on the current company
Before this commit, when editing a default value for a field, it was always setting
th default on the company of the user (as set on the user), not on the current company the user
is logged into.

This was largely misleading.

After this commit, we get and set the default on the current company.

Some work need to be done in master to improve those flows.

opw-4427384

closes odoo/enterprise#76819

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>