[FIX] account_bank_statement_import_camt: handle more multi-currency rate
The current implementation to handle the rate for multi-currency entries
in CAMT creates problems when fetching the source rate from the file:
the source rate can be taken from the SrcCcy or the TrgtCcy node, which
should be compared with the source currency and target currency
respectively. It is not done that way and is too restrictive (when the
SrcCcy is found, it is always compared to the target currency, which
leads to an incorrect transaction amount).

Moreover, when the currency conversion introduces a exchange loss/gain,
it is put back on the largest transaction of the entry.

opw-4393164

closes odoo/enterprise#76828

Signed-off-by: William André (wan) <wan@odoo.com>