[FIX] account_online_synchronization: Optimize fetching cron activation logic
This commit ensures the asynchronous fetching cron is activated or deactivated
only when necessary.
Previously, the cron was executed every 5 minutes to check for online accounts
to fetch transactions, even when none were present, leading to useless call to
the cron.

With this update, a toggle function is introduced, which is triggered whenever
an online account/link is created, modified (with the fetching status field),
or removed.
Before deactivating the cron, the function verifies that with don't have any
journals with an online account set on it, preventing unnecessary
execution.

This commit also removes some technical debts inserted when we try to fix the
cron issue on Odoo Online.

no task id

closes odoo/enterprise#76843

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>