[FIX] spreadsheet_dashboard_sale_subscription: Fix dashboard filter
The `MRR Evolution` dashboard contained a relation filter that would
wrongly refer to the model `sale.temporal.recurrence` even tough its
field matching pointed towards `sale.subscription.plan`.
From a setup POV, the filter was invalid but the only part we use
is the field matching, which was correct in itself. The issue came
when we wanted to edit the filter in a database which did not have
the model `sale.temporal.recurrence`.

task-4467133
opw-4432567

closes odoo/enterprise#78034

X-original-commit: 14bfd64cd0393b1e51fddb0d3c168295f987df25
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>