[IMP] l10n_uy_edi: Cancel doc if rejected by DGI
This commit aims to automatically set the status of a “Rejected”
document to “Cancelled” while preserving the sequence number assigned
by DGI (based on the CAE) to ensure all documents,
including rejected ones, remain in the system with their original sequence.
This is to maintain proper accounting records and provides visibility of
all rejected documents.

closes odoo/enterprise#80597

X-original-commit: 9d30a0ae06544cfaf04e84e67df388d99917390d
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>