[IMP] l10n_be_codabox: retrieve CODAs from 1 year ago
Accountants that start a new DB in Odoo may want to
fetch more than 3 months of historic for CODAs (i.e.
bank statements). So we're raising this to 1 year
so that it's enough time to fetch every transaction
needed for an annual closing.

Although this may seem like a lot to import, and we may
timeout, it's not really a problem because after each
imported CODA, we commit it. So next time the user
syncs the CODAs, we'll continue where we left off, and
we'll import only the CODAs whose date is after the last
imported one.

closes odoo/enterprise#76957

Task-id: none
X-original-commit: fd7a7f8e103aa411e439916fa5758de8a0d78516
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>