[FIX] l10n_fr_fec_import: Fix currency rounding
An issue may occur where moves are unbalanced when FEC file lines
contain a device with an imprecise rounding factor.

Steps to reproduce:
- Create or use a currency with rounding = 1.
- Add rounding to debit or credit in a FEC file, ensuring the rounded
unit amount is not equal; for example, two debit lines at 100.5 and one
credit line at 201 (see example in test).
- In the Idevise column of the file, use the name of the created device.
- Attempt to import the file.

An error will occur:
Moves report incorrect balances: Move with name '<name>' has a
balance of 1.00.

The issue arises because the created device is used to round the amount
instead of the company's device, resulting in 100.5 being rounded to
100, hence the two line of 100 no longer balances with 201.

opw-4397836

closes odoo/enterprise#79698

X-original-commit: b7b398217944eb32948abb0064383ec1eb88a9a6
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>