[FIX] web_studio: default value for selection radio
Since commit odoo/enterprise@d6d8aeff58e46373f72fc932cd84822a29ea76fb, the way to display
the default value input in Studio's sidebar was erroneous. In particular, A selection field with a radio widget associated had
its default value also displayed with a (buggy) radio input. Also, one2many fields were authorized to have a default value set, which was not correct.

This was due to the fact that we took the widget of a field to display such value, instead of just the field's type.
Also in the commit just mentioned, one2many fields appararently enabled the editing of default value, which was incoherent
because manymany fields did not enable that feature. For consistency, this commit fixes this by not enabling editing a o2m default value.

Work should be done in future versions to allow editing a default value for many2one fields and domain fields.

task-4285805

closes odoo/enterprise#76950

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>