[FIX] account_reports: allow moving more than 365 years
In accounting reports that use periods (like the tax report or profit
and loss), the date selector is broken when you try to go back or
forward for more than 365 years.

The reason is that it was computing the resulting year by looping over
the number of years to go back/forward, and each time adding or
subtracting the number of years as days. When over 365, it would go
back/forward more than 1 year per iteration.

This commit fixes the issue.

task-4410887

closes odoo/enterprise#77098

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>