[FIX] documents, documents_hr: fix company folder protection
How to reproduce:
- install documents_hr
- In Document, create a folder prjComp2 in /Marketing/Brand 2
- create a second company: SecondCompany
- Select SecondCompany
- Activate centralization of document for HR and select prjComp2 folder
- Log as Marc Demo and delete the folder Brand 2 (Move to Trash + delete)

There is no error while this folder should have been protected by the method
_raise_if_used_folder called when archiving or deleting a folder.

The cause is that Marc user doesn't have access to SecondCompany and as the
check is not done in sudo in _raise_if_used_folder, the folder is not
identified as a folder to be protected. We solve the issue by running the check
in sudo.

Technical note: the protection mechanism is defined in the documents app but
tested in documents_hr because there is no `documents.document` relational
field on `res.company` in documents.

Task-4480581

Part-of: odoo/enterprise#77162
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>