[FIX] l10n_ke_edi_oscu: make sure UoM code can be set on UoM
The problem right now is that if you use a new UoM, the
related Kenya eTIMS code is not in the category view.

Also, the error message you get, does not tell there is
a problem with the UoM in itself as it is not checked before
sending the item to eTIMS.

So, we add the code in the UoM category view and provide a
message when it is missing.

opw-4395902

closes odoo/enterprise#77116

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>