[FIX] documents,documents_project: fix document deletion
How to reproduce (Move to the Trash):
- Install documents_project
- Remove any access to the project folder (except to Mitchel Admin)
- Login as Marc Demo
- Upload a file in Marketing
- Move that file to the Trash

You get an access error while you should be able to move that document to the
Trash.

How to reproduce (Delete):
- install documents_project
- Login as Marc Demo
- Upload a file in Marketing
- Move it to the Trash
- Remove any access to the project folder (except to Mitchel Admin)
- As Marc Demo, go to the Trash and delete the document

You get an access error while you should be able to delete that document.

We solve the problem by doing the check in sudo (in
unlink_except_project_folder).

Technical note: We haven't done that test in documents (where there are already
delete tests) because we had to change the test so that they are executed after
all the modules are installed and that cause query count issue in performance
tests.

Task-4480581

closes odoo/enterprise#77162

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>