[FIX] account_sepa_direct_debit: relax SDD date constraint
By default, SEPA requires a 5 day notification period prior to
collecting a payment, if no payment has ever been collected using the
related mandate.

However, it could be that the company used another software before and
imported their existing mandates (that have no payments yet in Odoo).
In this case, the 5 day period does not apply and can be relaxes to the
default 2 day period for running mandates.

Instead of blocking the user when their required payment date is not
at least 5 days from today (on first-time mandates), we now only show a
warning at the top of the form, with links to the related payments.
Blocking the user will only happen if the required payment date is not
at least 2 days from today.

task-4472278

closes odoo/enterprise#77134

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>