[IMP] documents: make various improvements
This commits brings the following changes to Documents. Mostly UI
improvements:

Keep image preview aspect ratio
==============================

Currently, images are distorted because the aspect ratio is not respected.
Images must retain their original proportions, with the addition of
horizontal or vertical spacing if necessary, so that each so that each
preview frame has the same dimensions.

Also, we want the background of the image wrapper being the same as the
body.

Add tooltip to toggle chatter btn
=================================

Add the following tooltip to the toggle chatter button: "Info & tags".

Display existing email alias in help content
===========================================

When an email alias has been configured for a given folder, and a search
on this folder returns 0 record, the helper content displayed should
include the email alias.

In case of a search on 'Home' or 'Company': no message about alias, like
for 'My Drive' or 'Shared with me'.

Also, in 'documents_details_panel':
Change labels Email Alias for Email, and Alias Tags for Tags.

Enhance Home kanban view style
==============================

Enhance the style of the file / folder group titles by turning their
h3 element into a h4*, and removing the border-bottom class.

* actually we keep the elements as they are, to avoid conflicting with
potential customizations, and reduce the police size.

Avoid scrollbar on topbar Chatter
================================

Limit the display, on large devices, of the horizontal scrollbar in the
Chatter topbar by adding 20px to the Chatter container width, and
showing the scrollbar only when the cursor is over that area.

Scroll to selected record when info panel on
===========================================

When toggling the information panel on, the space available to display
documents/folders is significantly reduced. Sometimes, the selected
document moves further down the screen and disappears from the view.

This commit makes it so the view instantly scroll to the selected
document when the info panel is toggled on.

Improve email aliases display
============================

We remove the 'Email Aliases' menu item from the 'Configuration' menu
and add an optional column in the list view to display the email aliases
(widget=email).

Also we hide the Alias Name and Domain columns. We keep the fields in
the view as they are needed for the details panel.

Hide member invite instead of disabling it
=========================================

We currently disable the member invite component in the permission panel
for users with insufficient rights (not editor) or for inactive
documents/folders.

But the restriction is not really obvious (visually) for users. We
decided to hide the invitation to members to limit their confusion.

Add database creator to highest security group
==============================================

Generally the user who creates the database receives the highest permissions
for each application, except for specific cases.

This commits follows the general logic and links 'base.user_admin' to
the 'System Administrator' group.

Fix spacing in control panel
============================

This commit aims to limit the vertical spacing between the control
panel and the menu section.

The minimum height applied to breadcrumbs to prevent vertical shift
of the view when the breadcrumbs changes dimension has been modified.
This height is defined in 'rem' rather than 'vh' to avoid affecting the
height of the control panel, particularly on high-resolution screens.

task-4444059

closes odoo/enterprise#77554

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>