[FIX] web_studio: find the right record when adding fields in header/footer
Before this commit, when in the report editor, trying to add a field in the footer,
the main doc's fields were proposed.

This was wrong because header and footer are generic views that can treat any record at once.
They just set an internal variable for the company of the record, which should be the only available model to add
fields from.

After this commit, it is not possible to add fields from anything else than the "company" variable

opw-4471868

closes odoo/enterprise#77234

Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>