[FIX] documents: clean portal action
We disable action that portal user cannot perform:

In the preview:
- hide Split Pdf in the preview
In the cog menu:
- remove "Move to Trash" as portal user can't archive document (and can't
create folder nor access to the trash anyway)
- Info&Tags as it opens the chatter which is disabled for portal user
In action:
- Info&Tags as it opens the chatter which is disabled for portal user
- Duplicate only present if it doesn't lead to an error
(canDuplicateSelection method). We also remove the line separator in the action
if the user is not internal otherwise it leads sometimes to an empty section.
In the "New" drop down button:
- hide "Folder" (create Folder)
- hide "Link" (create link):
fails because of (users | folders.owner_id).fetch(['partner_id']) in
_prepare_create_values (no access to user)

Hide share button for portal user: Although the portal users could share any
documents they have access to (by sending them by email for example), we remove
the share action (for document and folder) for them as the only thing they
could do was to change the rights for user that had already a link to that
document.

Task-4221258

closes odoo/enterprise#77281

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>