[IMP] {website_}knowledge: move html_field migration to html_editor
Removing the Excalidraw command means that the html migration system must be
moved to `html_editor`.
This commit thus removes the sytem from Knowledge and adapts the manifests to
take into account the move done, and updates a test to account for the new html
editor version (1.1).

task-4489076

closes odoo/enterprise#77321

Related: odoo/odoo#194066
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>