[FIX] hr_payroll_account: Set default company currency for payslip payments
Currently, an error was occurring when a user tried to reverse a journal entry
by following these steps:
- Install 'Accounting' and 'Payroll' module
- Enabled payroll entries in Payroll > Configuration > Settings.
- Created a new employee and its corresponding contract in
  payroll > Create a bank account and enable "Send Money" in
 'Private Information'
- In 'Contract' change state from 'New' to 'Running' > Click on 'Payslips'
- Computed the payslip > generated draft journal entries > add lines in journal
  items and click 'Pay' in payslip.
- Go to 'Journal Entry' and make a 'Reverse Entry'.

ValueError: Expected singleton: res.currency()

The issue occurs at [1] because, when a user makes a payment from a payslip
after posting journal entries, the 'currency_id' is not available.

[1] - https://github.com/odoo/odoo/blob/0583db332efdc884975da1a074c91981d2865c9d/addons/account/wizard/account_payment_register.py#L972

This commit will fix the above issue by providing default company currency when
currency is not available.

sentry-6207053954

closes odoo/enterprise#77334

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>