[FIX] project_mrp_workorder_account : fix failing test
This commit's purpose is to make the profitability test works without
demo data. The issue is that without some specific groups, the
'workorder_ids' field is not visible inside the views, making the
creation of record impossible. Adding these groups in the config of the
test fixes the issue

version 18.0-master

closes odoo/enterprise#77350

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>