[FIX] account_sepa_direct_debit: unlink mandate and invoice on payment deletion
Steps to reproduce:
- Select an EU company and provide a SEPA creditor identifier (for example BE55ZZZ0455530509)
- Have a contact with a European IBAN
- Have a company bank journal with a European IBAN
- Create a valid Direct Debit Mandate for your contact
- Create an invoice for that contact and register a payment using SEPA
- Now delete that Sepa payment from Accounting > Customers > Payments
- Open the customer's Direct Debit Mandate > It shows one invoice was paid using that mandate although the payment was deleted
- Expected: link between invoice and mandate should not be kept if payment was deleted

Cause:
The field `sdd_mandate_id` of the invoice is not reset when the payment is reset to draft.

Solution:
Add an `api.ondelete` method on `account.partial.reconcile` which checks if any of the unreconciled moves is an invoice with `sdd_mandate_id`. In that case, and if the unreconciled payment is the only SEPA payment, set `sdd_mandate_id` to False.

opw-4277982

closes odoo/enterprise#77475

X-original-commit: 933f6950f5cb0b6737ce39c7f672efc5e82069f0
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>