[FIX] account_asset: allow to create asset models
Steps to reproduce:
- Install Accounting
- Go to "Accounting / Configuration / Management / Asset Models"
- Create an asset model

Issue:
Upon save, the following UserError is raised:
"You cannot add or remove bills when the asset is already running
or closed."

Cause:
A check has been added to prevent adding bills to a running asset.
The check is excluding draft assets but not asset models that are
assets in "model" state.

Issue:
Exclude asset models from the check.

opw-4479698

closes odoo/enterprise#77576

X-original-commit: 69b28ef51d7c3b8341ead144957a578f137b82fd
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>