[FIX] account_budget: use price_subtotal in committed amount
Steps to reproduce the issue:

1. Create an clean Analytic Account and use it in a Budget
2. Create a Purchase Order with Analytic Account in the Distribution and a Discount then Confirm
3. In the Budget report, the Committed Amount does not include the Discount

Explanation:

In the `purchase.order.line` side of the calculation of the committed amount of the budget report, we are using `unit_price` as base, because of it having a `not null` constraint, then adding a lot of other variables to extract the correct amount: `currency_rate`, `analytic_distribution`, `product_qty` and `qty_invoiced`. This does not include `discount` nor price-included `taxes_id`.

Fix reasoning:

To avoid an issue regarding down payment lines, which have no `price_subtotal`, we will prioritize the field with a fallback on `price_unit`.

opw-4391098

closes odoo/enterprise#78167

X-original-commit: ac95249993777d70f5203e2440c5c21ed41faa6e
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>