[FIX] account_reports: improve xlsx styling
After the [indentation fix] for xlsx exports, the styling logic was
simplified to always show levels 0, 1 and 2 in bold and all deeper
levels not in bold. Before that, there was a separate rule that defined
that for any line having `caret_options`, we wouldn't use bold.

Because of this, reports like the general ledger would display all
detail lines indicating journal entries in bold (since they are in
level 2). This looks very crowded and not very readable at all.

In this commit, we improve the styling rules to always disable bold
on the deepest line level. That way we can never have a XLSX file with
all lines in bold, and it is much easier to read the file.

Meanwhile, the style management was refactored a bit to make it simpler.

opw-4501822
opw-4458752

[indentation fix]: https://github.com/odoo/enterprise/commit/9246a864f91e7facd06fa630fa0d23a0323ce497

closes odoo/enterprise#79131

X-original-commit: 03ffec66abf5a958310221ec00bdf746f679ae09
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>