[IMP] documents: less rpc calls to set up automation window action
Automation rule trigger field ids can now be deducted from a default
filter domain passed in the context of an action. As a result, in the
action used to direct a user of the Documents application to the
Automation Rules, it is no longer necessary to pass the trigger field id
to be displayed in the new rule creation form.

Only the filter domain is required to achieve the same result. This
makes it possible to delete an orm call without altering the implemented
feature.

see odoo/odoo#193576
see odoo/enterprise#75719
task-4502214

closes odoo/enterprise#77515

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>