[FIX] documents_spreadsheet: drop spreadsheet auto-clean
The current auto-clean feature doesn't work properly and
it deletes spreadsheet that shouldn't be deleted.

1) copy a spreadsheet and leave it (no update)
2) upload an XLSX file (see commit 37e279b02ab4)

Given that it's been broken several times and the usefulness of the feature
is debatable, let's just drop the feature. No feature, no bug 🙃

If users create empty spreadsheet, so be it. They'll clean
their mess themselves.

closes odoo/enterprise#77573

Task: 4506185
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>