[FIX] spreadsheet_edition: Fix undo/redo and datasources reload
The current implementation of spreadsheet history does not support
UNDO/REDO commands as those were never designed to be rollbacked in the
first place (to rollback and UNDO, you cast a REDO).

Furthermore, the datasources are not properly reloaded when navigating
the history. When selecting a revision for which the domain or more
generally the definition of datasource is altered, the latter is not
reloaded and therefore the values displayed do not correspond to the
definition in place.

This revision changes the flow by simply re-instanciating a new `Model`
every time we change the target revision.

Task-4506832

closes odoo/enterprise#81302

X-original-commit: 9a3d3bebed585f172adb522f7a25b60baf92f4e1
Related: odoo/odoo#201601
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>