[FIX] account_base_import, l10n_fr_fec_import: FEC import button display
Issue
-----
The "Import FEC" button is only visible when in a company set in France or
Luxemburg, despite the localization being installed.

Steps to reproduce
-----
- Install Accounting & the French accounting localization
- Settings > Accounting > Import

Discussion
-----
By default, when the FEC module is not installed, there is a message prompting
the user to install it with a link. This message is only visible when in a
French or Luxembourgian company.

Ticket:
opw-4356726

closes odoo/enterprise#77908

X-original-commit: af68b70e5356e225425b0b167b950a08de92e6f1
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>