[FIX] account_asset: print unfolded lines in depreciation schedule
If "Unfold All" is not selected and "Group by Account" is,
printing the Depreciation Schedule won't print the unfolded lines.

Furthermore, if you unfold some lines and then refresh, all the lines
are refolded.

This behaviour is different than the other reports, where the lines you
unfold are printed and remembered during a refresh.

The fix is to mark the parent lines generated by `group_by_account` as
unfolded if the line is present in `options['unfolded_lines']`.

task-4507580

closes odoo/enterprise#78074

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>