[IMP] documents: keep permission when replaced as owner
When the ownership of a document changes, the previous
owner can potentially lose access to the document completely,
which was perhaps not desired. If the user adds the previous
owner to the members at that point, this will cascade to the
children, and was perhaps not desired either, so it could
require updating the children afterward too...

We here replicate the logic used in pinning to company root
where the previous owner is added as editor.

Task-4510838

closes odoo/enterprise#77783

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>