[FIX] planning: round full time required hours to the minute
This commit aims to ensure that the hours for flexible resources are not rounded when displayed on the Gantt progress bar. Previously, the hours were being rounded to the hours, which could lead to inaccuracies in the representation of the expected working hours on the gantt view. By changing the rounding to the minutes, we ensure that the Gantt progress bar accurately reflects the precise hours worked by flexible resources.

Steps to reproduce:
-------------------
1. Create an employee with flexible hours ex. 42:30h per week, 8:30h per day.
2. Create a shift for the same employee on planning.
3. In the gantt view, when setting the date range to a week, the expected working hours of the period is showing 42h instead of 42:30h

closes odoo/enterprise#77782

Task-id: 4506633
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>