[FIX] mrp_mps: actual demand in multi-steps
If a warehouse has multi-steps deliveries, the actual demand of a
product only takes the real outgoing move into account: the move
that goes directly to the customer location.

Since moves are not created in advance anymore, we need to take
the earlier moves in the chain into account instead while making
sure that we only use a single move at a time.

The move used will be the last created in the chain, the one without
`move_dest_ids`.

task 4510137

closes odoo/enterprise#77787

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>