[FIX] hr_payroll: broken tour on new Chrome
Since version Chrome 128, the "new" headless mode has been set as
default but the `payroll_dashboard_ui_tour` failed on a step relying on
the "blur" event to update the Todo's name.

After very long and quite thorough research what was discovered are the
following:
- this issue happens only once the `website` module has been installed
- the tour is started from "/" (in Python) but actually run from "/web"
  (in JS), resulting in a useless first load.
- the "blur" event triggered by the `text_run ...` tour's command never
  reaches the `t-on-blur` OWL's handler, while no listener that may have
  "eaten" the event has been found...
- making the "blur" event bubble (which is not its normal behavior) does
  have some impact on the issue...
- running the tour directly from the backend (e.g. `start_tour("/web",
  ...)` doesn't present this issue.

While the exact root cause of this issue hasn't been identified so fare,
this commit sets the tour to be run directly from the backend (which
makes sense as it is a backend-only test tour anyway) as a
workaround.

Note: while being an issue in automated test, the feature operated by an
actual user looks to be working properly.

closes odoo/enterprise#78007

X-original-commit: a07631c9c58ac52866bbeb31fcf2653267190618
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>