[FIX] account_reports: link to parent tax closing
When you had tax closing entries that depended on a main closing entry,
the posting of that main entry would log a message on each depending
entry containing a link to the main closing entry.

In [this commit] the way that message is constructed was changed,
making it safer but also causing the link HTML to be escaped. It would
just show raw HTML in the logged message instead of a link.

This commit constructs the message in a correct and safe way so the
link works again in the logged message.

[this commit]: https://github.com/odoo/enterprise/commit/09f5b35493bee7b63e96e8b28ad741aeb5ee80e5

closes odoo/enterprise#77826

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>