[FIX] *_stock_*: pass `previous_product_uom_qty` as kwarg
*: industry_fsm_stock, sale_stock_renting, sale_subscription_stock

The override of `_action_launch_stock_rule` was passing its
`previous_product_uom_qty` keyword argument as a positional parameter to
the `super` call, preventing other overrides from passing it with its
key name when the argument was packed with the `**` operator.

This commit ensures that the argument is now passed correctly with its
key name. It also makes the overrides kwargs-agnostic by packing and
unpacking them with `**`, thus preventing future issues related to
signature changes in the parent method.

Part-of: odoo/enterprise#77839
Related: odoo/odoo#193457
Related: odoo/documentation#11823
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>