[FIX] account_sepa: allow ISO 20022 when no IBAN bank account
We already support the non-SEPA ISO 20022 method in the `account_sepa`
module. However, we were still requiring the origin bank account to be
an IBAN account, according to SEPA rules.

However, since ISO 20022 can be used outside of the SEPA zone, we don't
need to check for an IBAN in that case. This commit fixes that, so non
IBAN bank accounts can use ISO 20022 payments.

closes odoo/enterprise#78023

X-original-commit: cac7cede1bb0e37a511e62c60d308fcba2e7169c
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>