[FIX] account_loans: prevent memory error in loan creation
Description:
When creating a loan, users encounter a memory error if they have used
accounting. This occurs because the `currency_id` field is related to
the company's `currency_id` and is not readonly, causing the ORM to
perform an inverse related and update the company's `currency_id`.
This is problematic since changing the currency triggers a massive
recompute of nearly all accounting records unnecessarily. For example,
in the case of this customer, it was trying to update
191219 account.analytic.line, which caused this memory error.

The initial design aimed to show the monetary symbol on the form when
the `onchange` in `web/models/models.py` was triggered, but this
approach led to the issue.

Fix:
By adding `company_id` to the view as an invisible field, the onchange
can retrieve it to display the monetary symbol without triggering the
problem. This change is backward-compatible; customers who haven't
updated their views will only miss the monetary symbols on the form
view until the record is saved.

Reference:
opw-4486739

closes odoo/enterprise#77990

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>