[FIX] web_enterprise: fix wrong color matching in dark mode
This commit fixes an issue about the `$o-color-original` color map values
sometimes not matching the light mode one.

When we introduced the redesign of the dark mode with Milk (`saas-16.3`),
we fine-tuned the badges design in dark mode by slightly tweaking some
colors from the `$o-color-original` map, which contains the colors that
will be used on badges (as well as other UI elements).

While most of the colors were normally tweaked, some colors received a
new hue value, meaning the color would not match the light mode one.

As this is inconsistent and unexpected, we review the colors to match
the light mode ones.

task-4522539

closes odoo/enterprise#78046

X-original-commit: bc0450531aec11e5e8c45136180a562d8ed6fc6d
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>