[FIX] helpdesk_stock: Split by partner for suitable products
Description:
Following odoo/enterprise@c9ceba2cf5982c244fe1ad3054d1463d9eca0e8c,
the set of suitable_partners used to determine which products can be
associated with a ticket is based on the `commercial_partner_id` of
`self`. This works well for form views, where `self` is a singleton.
However, since the field `product_id` is present in the list view, which
has a domain depending on `suitable_product_ids`, the computation
is triggered for all tickets in the list view, making `self` not a
singleton.
This causes the set of suitable partners to be larger than necessary, as
it becomes the collective set for *all* tickets referenced in `self`,
leading to false positives where products associated with a specific
partner's sales/delivery history might appear suitable for another
ticket with a different partner.

To address this, we split `self` by `commercial_partner`, as the
set of products that could be suitable depends on the partner's
sales/delivery history, not on individual tickets.

Reference:
opw-4444078

closes odoo/enterprise#78550

X-original-commit: 48cf763985bcd4bff56dcdef3cc07a805815e126
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>