[FIX] point_of_sale, pos_settle_due: prevent creation of empty orders
Before this commit, depositing money for a customer with a zero balance
could result in the creation of an empty order upon refreshing the
browser. This issue arose because empty orders without payments were
being synchronized, leading to unintended empty orders in the system.

This commit addresses the problem by ensuring that empty orders without
payments are not synchronized and prevents the creation of such empty
orders during the deposit process.

opw-4483049

closes odoo/enterprise#78217

X-original-commit: b5ad5dbc4325174c0d944e84f2b8cc6adc859f6a
Related: odoo/odoo#195850
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>