[FIX] appointment: fix the access error when accessing internal meetings
How to reproduce:
=====================
- ​As Mitchel Admin, create an appointment type, keep it unpublished
- Send an invite (with token) to Demo
- Log in as Demo (User access level)
- Book a meeting with your manager
- You get a calendar event in your agenda
- Try to open its form

Before this commit:
=====================
It gives access error because the user does not have
permission to read the appointment type.

Technical reason:
=====================
The record rules only grant access if the appointment is scheduled
based on resources or if the user is part of the staff.

After this commit:
=====================
Users will be able to access the appointment type because sudo will be added.

Task-4526059

closes odoo/enterprise#78435

X-original-commit: e52b13c9fc9cb309351f527bfbe961fcb0fb961e
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Nihar Patel (nihp) <nihp@odoo.com>