[FIX] website_documents: avoid cascading website deletion
Due to an oversight when migrating documents share to
documents in Sharepocalypse, documents were being
deleted when their related website was deleted.

This is not a desired behavior.

Even though setting the "correct" website is possible
(and desired) when there is a `company_id`, there is no
problem with having empty `website_id` in the other case.

opw-4494091

closes odoo/enterprise#78163

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>