[FIX] l10n_cl_edi[_stock]: Error when validating Boleta
In Chile localization it is possible to add a Related Document when
creating an Invoice Document. These documents info are then added to the Electronic
Invoice sent to the government.

Currently, when creating a Boleta Electronica and adding a Cross
Reference Document, the EDI system will not validate correctly the xml,
with error
`Error en Schema [0] LSX-00204: extra data at end of complex element`

This occurs because the system first creates a dte from a common xml schema, then
wrap it with the appropriate envelope.
However the schema of Boletas slighly differs[1], as it doesn't have the
field FchRef, so it gets rejected by SII

A solution is to modify the code of dte_subtemplate to make sure that FchRef
only appears if the Schema corresponds to http://www.sii.cl/SiiDte EnvioDTE_v10.xsd,
and not to http://www.sii.cl/SiiDte EnvioBOLETA_v11.xsd

[1] https://www.sii.cl/factura_electronica/factura_mercado/formato_boletas_elec_202412.pdf

opw-4433202

closes odoo/enterprise#78177

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>