[FIX] sale_amazon: don't send picking information when Amazon it
In some cases, Amazon can confirm the delivery of a FBM order before the
user confirm the picking. In these cases, 2 problems arise:
- The synchronization of the picking will fail because Amazon will
return an error, since the picking was already delivered.
- If the user doesn't fill the Amazon required fields, there will be a
user error, even if the carrier reference and tracking numbers are
already available in Amazon's backend.
To solve this, as we know when an order was successfully delivered, we
won't consider these pickings as being needed for Amazon any more.

task-4524195

closes odoo/enterprise#82638

X-original-commit: a6bad2428f72b6b6cdd967049254b6d1016b6365
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>