[FIX] account_reports: fix report with sections using custom action
Steps to reproduce:
- Install 'l10n_es' and switch to Spanish company
- Go in Accounting > Tax Report and select the report "Mod 390"
- Click on the button to download the BOE
- Fill the informations in the wizard
- Click on generate BOE
- An error pops up

Cause:
The report Mod 390 is composed of several sections. The method `dispatch_report_action` is called on the first section instead of the entire report. As the first section has no custom handler, an error is raised when trying to read on it.

When exporting a report wih sections, `dispatch_report_action` is called on the first section, it detects that this is a section and calls the same method on the parent report.
This commit (https://github.com/odoo/enterprise/commit/8213813e68f02d88934285a44a3413505f6ca187) removed the line getting the options for the parent report call to keep the specific options of each section.
Specifically the "report_id" key in the options is the id of the first section, causing the method to be called on the

Solution:
Setting the key "report_id" in the options of the parent report.

opw-4501376

closes odoo/enterprise#78185

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>