[FIX] l10n_de_pos_res_cert: prevent IndexError when splitting orders
Before this commit, in the restaurant, splitting an order could result
in an `IndexError: list index out of range`. This issue occurred due
to incorrect handling of order lines during the split process.

This fix ensures proper validation and handling of order lines to
prevent such errors, improving the stability of the POS system.

Steps to reproduce:
- Create an order
- Refresh browser
- Attempt to split the order.
- Observe the `IndexError: list index out of range` traceback.

opw-4451836

closes odoo/enterprise#78184

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>