[IMP] account_online_synchronization: delete user
This commit will authorize to delete an online synchronization from the
cog menu even if the connection is in error. Previously, we added a
check that if the connection throws a user error or a redirect warning
we can't delete it. We will
remove this.

Backport of odoo/enterprise#72902

task-4526918

closes odoo/enterprise#78245

X-original-commit: 498daf54b2abdbe8b1c36c4ba2a8222107b0cf1a
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>