[IMP] documents{_spreadsheet}: make a few improvements on share panel
Clarify owner on share panel:
============================

The panel must still indicate the owner (except for Odoobot)
of the folder/document, when the partner list with access is empty.

In addition, we don't want the owner to appear twice in this list, and
there's no point in displaying his role. It may happen because someone
sets the document's owner permission using the invitation member component
of the panel. To avoid this, the owner can no longer be found from this
component.

Access to permission panel data:
===============================

As portals users can't use the share panel, access to the permission
panel data is restricted to internal users.

Editing permission expiration date:
==================================

We don't want users with limited read access to a document/folder to be
able to edit/remove their expiration date.

task-4527397

closes odoo/enterprise#78199

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>