[FIX] stock_barcode: fetch relevant archived records in lazy barcode cache
Issue:

The barcode relies heavily on the data fetched in its LazyCache to work
properly (in order to avoid rpc call's each time records data's are
needed). However, some of these data's are loaded using search methods
that ignore archived records by default. This leads to the creation of
a cache that do not contains records that are expected to be there and
hence will lead to tracebacks.

Steps to reproduce:

- In the settings enable "Multi-Step Routes"
- Inventory > Configuration > Warehouse Management > Locations
- Create a new location WH/Stock/Love and a children location
  WH/Stock/Love/Layer
- Create and confirm an internal transfer from WH/Stock to WH/Stock/Love
  for 1 unit of any storable product in stock
- Change the location dest of the move line to WH/Stock/Love/Layer.
- Archive the location WH/Stock/Love
- Open your internal transfer in the barcode App.
> Traceback: An error occured in the owl lifecycle... Record
stock.location with id=... doesn't exist in the cache, it should return
by the server at LazyBarcodeCache.getRecord.

Cause of the issue:

When entering the barcode man component, a call of the `get_barcode_data`
is launched to fetch and populate the barcode data's and data are added
to the barcode lazy cache in order to retreive record values without
additional rpc calls:
https://github.com/odoo/enterprise/blob/cc21959226e0bdb6b840f5023f08b2fdc0590023/stock_barcode/static/src/components/main.js#L76-L93
https://github.com/odoo/enterprise/blob/cc21959226e0bdb6b840f5023f08b2fdc0590023/stock_barcode/static/src/models/barcode_model.js#L32-L34
For instance, when redering the picking template, the location dest id
will be retreived from the cache:
https://github.com/odoo/enterprise/blob/cc21959226e0bdb6b840f5023f08b2fdc0590023/stock_barcode/static/src/components/line.js#L7-L9
https://github.com/odoo/enterprise/blob/cc21959226e0bdb6b840f5023f08b2fdc0590023/stock_barcode/static/src/models/barcode_picking_model.js#L1077-L1079
The traceback occurs because the record was actually never added to the
data's and hence never added to the cache because part of the
`get_barcode_data` call is performed with searchs that ignore completely
the archived records (unless the 'active_test' key is in the context):
https://github.com/odoo/enterprise/blob/cc21959226e0bdb6b840f5023f08b2fdc0590023/stock_barcode/models/stock_picking.py#L102-L106

opw-4423110

closes odoo/enterprise#79194

X-original-commit: e783c62b6710e29cb3e242be152a4ae02bdf25b1
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>