[FIX] documents: fix control panel wrong display
When you reduce the screen width or view the page on devices with small
screens, the control panel button group rearranges itself in an
unsightly way (e.g. excessive button height).

For mobile devices, we use overflow auto on the action btns
of the panel. Then they keep their aspects and are still accessible to
users.

task-4471881

closes odoo/enterprise#77039

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>