[FIX] account_reports: multi generated deferrals for non-past period
Steps to reproduce:
1. Deferred mode = Manually & Grouped
2. Create a deferred from 1/1/25 -> 12/31/25
3. Go on the deferred report
4. Select the current month or a later month than today
5. Generate the deferrals
6. Click again on Generate the deferrals -> Permitted -> Not OK

This happens because when we generate the deferrals for
a current of future period, the deferrals are set to be posted on a future date. Hence, they're still in draft.

So the `filter_already_generated` wasn't "seeing" that
those draft entries were already created, and thus one
could click the button many times.

task-4497852

Part-of: odoo/enterprise#77738
Signed-off-by: William André (wan) <wan@odoo.com>