[IMP] account_reports: reverse warning banner in deferred
Currently, in the manually & grouped mode of the deferred,
if the user generate the deferrals for a period and then
reopens that period, a banner would warn the user that
the deferrals have already been generated for this period.
However, when the deferrals weren't yet generated, nothing was
shown on the report.

With this commit, we reverse the logic: we show the warning
banner if the user has not yet (completely)  generated the
deferrals for the current period making it easier for the
user to understand that something should be done. When the
deferred have already been generated, we keep an info banner
which allows him/her to see the previously generated deferrals

task-4497852

Part-of: odoo/enterprise#77738
Signed-off-by: William André (wan) <wan@odoo.com>