[FIX] account_reports: Make the fiscal year filter consistent with spreadsheet
Problem
---------
Currently, there is a discrepency between the year filter in the reports
and spreadsheet.

In the reports, when using the YEAR filter, it will use the fiscal year
that ENDS in the set year. On the contrary, in spreadsheet, the filter
will use the fiscal year that STARTS in the set year.

For example, if the fiscal year starts on the 1st of April (01/04),
- in the reports, the 2024 date range filters the moves between
  01/04/2023 and 31/03/2024;
- in spreadsheet, the 2024 date range filters the moves between
  01/04/2024 and 31/03/2025.

This behavior is explained by the fact that we use the current day to
compute the fiscal year.

Objective
---------
Make the report year filter behave like in spreadsheet.

Solution
---------
Update the initiation of the date option in the report options so that,
if the year filter is used, we compute the start of fiscal year that
will begin in the selected year rather than the fiscal year we currently
are in. We then use this fiscal year value to compute the relevant date
range like before.

opw-4493556

closes odoo/enterprise#79090

X-original-commit: 0c0af29f54fe94bee263d4052959cb6269c1360c
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>