[FIX] web_enterprise: make the db name visible in dark mode
Previously after activating debug mode and then dark mode,
the letters of database name are not visible clearly. It is
happening because in new bootstrap version we are having an
inappropriate color for mark tag in dark mode.

After this commit the database name will be clearly visible
in dark mode.

Task- 4389154

closes odoo/enterprise#78433

X-original-commit: ba7695f9dc8be917043034ce3246b06fbd6d5f31
Signed-off-by: Romain Estievenart (res) <res@odoo.com>