[FIX] web_enterprise: fix the properties icons colors in dark mode
Bug
===
The icon in the definition popover are currently black in dark mode, and it should be
light gray. The reason is that the component that it uses changed ( odoo/odoo@ea5c57ac16accb9e3499dfe160b0b27e1e68716d ),
but the CSS in the properties component wasn't updated (we now use a more permissive
rule to avoid future similar issues).

Task-4543117

closes odoo/enterprise#78526

X-original-commit: 83ffd9dc2ae88932dc98b9a1604a3a9b1dfe91ea
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>