[FIX] l10n_cz_reports_2025: Fixed missing taxes issue
Bug repro steps:
1. Install `l10n_cz_reports` module without `l10n_cz_reports_2025`
2. Delete some of l10n_cz taxes
3. Attempt to install `l10n_cz_reprots_2025` module
4. The module installation fails because it attempts to add reverse charge to non-existing taxes

The commit also makes the tax report XML export button redirect the user to install the `l10n_cz_reports_2025` module in case it was missing.
This is meant to replace the auto-installation of the module on the export button click, as it was not working.

Part-of: odoo/enterprise#78410
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>