[FIX] l10n_cz_reports_2025: Fix line name in control statement
In the Czech Control Statement, parts A1 and B1 are grouped by move id
and by the supplies code. Since commit #78145, account_reports handles
the use of selection fields in groupbys, leading to 'Undefined' being
shown in the expanded lines of the control statement.

This fix removes the workaround in the control statement, relying
directly on the account_reports approach to show a human-readable
line name.

closes odoo/enterprise#78410

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>