[FIX] sale_stock_renting: Make rental_loc_id visible in res_company view
Problem

Before this commit, if the customer deletes the rental location they would be blocked from creating a rental order as it is necessary for replenishment rules. The issue becomes critical because, once deleted, the rental location cannot be set to the company functionally given the field is not present in the view.

Solution

To fix this, this commit adds Rental Settings making the rental_loc_id field visible in the res_company view, allowing customers to select their desired rental location and resolve the issue independently.

- opw-3973957
- opw-4488182
- opw-4545985

closes odoo/enterprise#79785

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>