[FIX] hr_(payroll,recruitment_integration_monster): fix tracebacks in preparing_monster_data
Currently, we are getting an attribute error while executing the
below lines of code.

https://github.com/odoo/enterprise/blob/a0215115c13d5e32f1ba3c821f138b188083e76b/hr_recruitment_integration_monster/wizard/hr_recruitment_post.py#L16

https://github.com/odoo/enterprise/blob/a0215115c13d5e32f1ba3c821f138b188083e76b/hr_recruitment_integration_monster/wizard/hr_recruitment_post.py#L31

Error:-
"AttributeError: 'resource.calendar' object has no attribute 'hours_per_week'"

This is because the `hours_per_week`, `is_fulltime` is defined in `hr_payroll`.
But it is used in `hr_recruitment_integration_monster`. Which is not
dependent directly/indirectly on `hr_payroll`.

So this will lead to the above traceback.

We can resolve this issue by moving the hr_payroll code of
calculating the `hours_per_week` and `is_fulltime` to hr module.

sentry-6267814511

closes odoo/enterprise#78521

Related: odoo/odoo#196427
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>