[FIX] account_{accountant,reports}: deferred full_months with End of Month
Steps to reproduce:
1. Use the Full Months option in the deferred
2. Create a deferred o $1200 from 2023-01-31 -> 2024-01-30
3. Open the deferred report on March
4. We have an amount of 1000 in the Later which is incorrect

After March, we have April -> December ($100/month) so
we should have $900 instead.

This is because we were using comparisons with
ends of month in the `full_months` method.

By simplifying the code for the period boundaries,
we also simplify how the `full_months` method is
done which is simply by taking the difference of
the dates without taking into account the days
as any started month is fully accounted for.

task-4547202

closes odoo/enterprise#78487

Signed-off-by: William André (wan) <wan@odoo.com>
Co-authored-by: william-andre <wan@odoo.com>