[FIX] sale_commission: Remove display of bad aggregation in group by
Steps to reproduce:
-Make multiple sales and assign a salesman to it in order to have a commission
-Go to My Commissions
-Add a Group By
-The Achieved Rate column is displaying the sum of the achieved rates but it shouldn't display anything

Problem:
When making a group by on the commissions view and enabling the optional column achieved rate, the result of this group by would be the sum of all the achieved rate for this group, which is not correct.
As it was decided that it made no sense, the total would not be displayed anymore.
https://github.com/odoo/enterprise/blob/a0215115c13d5e32f1ba3c821f138b188083e76b/sale_commission/report/commission_report.py#L19

Fix:
The aggregator parameter was used to prevent the total to be displayed.

opw-4512739

closes odoo/enterprise#78537

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>